% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/answer.R
\name{answer}
\alias{answer}
\title{Simulate response(s)}
\usage{
answer(person, test, indeces = NULL)
}
\arguments{
\item{person}{}

\item{test}{}

\item{indeces}{If not NULL, answer questions with the given indeces, and update the person object.}
}
\value{
vector responses, or updated person object if indeces is set.
}
\description{
Given a person and (a subset of) an itembank, simulate a response pattern.
}

