% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{as.scalar2}
\alias{as.scalar2}
\title{Return as a scalar}
\usage{
as.scalar2(obj)
}
\arguments{
\item{obj}{Any object.}
}
\value{
\code{obj} as a scalar; \code{NA} if \code{obj} is \code{NULL}.
}
\description{
Same as \code{opencpu.encode::as.scalar}, except that it also turns \code{NULL} into \code{NA}.
}

