% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_item.R
\name{best_item}
\alias{best_item}
\title{Next item in an adaptive test. Takes a person and test object, and returns the index of the next item based}
\usage{
best_item(person, test)
}
\arguments{
\item{person}{person object}

\item{test}{test object}
}
\value{
integer item index
}
\description{
This function is a wrapper that sends the actual work to the correct subroutines.
}

