% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{constraints_correct_names}
\alias{constraints_correct_names}
\title{Check whether the name elements within the constraints elements are correct}
\usage{
constraints_correct_names(constraints, characteristics)
}
\arguments{
\item{constraints}{List of constraints}

\item{characteristics}{Data frame with characteristics}
}
\value{
TRUE if the name elements in constraints are correct, FALSE otherwise
}
\description{
Check whether the name elements within the constraints elements are correct
}

