% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{constraints_correct_operators}
\alias{constraints_correct_operators}
\title{Check whether the operator elements within the constraints elements are correct}
\usage{
constraints_correct_operators(constraints)
}
\arguments{
\item{constraints}{List of constraints}
}
\value{
TRUE if the operator elements in constraints are correct, FALSE otherwise
}
\description{
Check whether the operator elements within the constraints elements are correct
}

