% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{constraints_correct_structure}
\alias{constraints_correct_structure}
\title{Check whether elements of constraints have correct structure}
\usage{
constraints_correct_structure(constraints)
}
\arguments{
\item{constraints}{List of constraints.}
}
\value{
TRUE if elements are lists of length three with correct names, FALSE otherwise
}
\description{
Check whether elements of constraints have correct structure
}

