% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/constraintsClass.R
\name{createConstraints}
\alias{createConstraints}
\title{Convenience wrapper to add constraitns to a test object for use with ShadowCAT.}
\usage{
createConstraints(test, characteristics = NULL, constraints = NULL)
}
\arguments{
\item{test}{ShadowCAT test object}

\item{characteristics}{data.frame with characteristics, one row per item, one column per characteristic.}

\item{constraints}{list of constraints, each constraint is a list with three named values;
\code{name} the column name of the characteristic this constraint applies to. For categorical characteristics the level should be specified as \code{name/value}.
\code{op} the logoical operator to be used. Valid options are "<", "=", ">" and "><".
\code{target} the target value, numeric. If the operator is "<>", this should be a length two vector in between which the target should fall.}
}
\value{
test ShadowCAT test object, including the constraints set.
}
\description{
Convenience wrapper to add constraitns to a test object for use with ShadowCAT.
}

