% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exposure.R
\name{createExposureConstraint}
\alias{createExposureConstraint}
\title{exposure control}
\usage{
createExposureConstraint(test, exposure, eligible, feasible, total, target)
}
\arguments{
\item{test}{}

\item{exposure}{vector of actual exposure counts.}

\item{eligible}{vector of eligibility counts (including lifting of restrictions.)}

\item{feasible}{count of respondents for whom a feasible solution was found.}

\item{total}{count of total respondents so far.}

\item{target}{(vector) target exposure rate(s). If scalar, will be used for all.}
}
\value{
test
}
\description{
Sympson-Hetter style exposure control. Van der Linden & Veldkamp, 2004.
}
\details{
This is half-implemented. The routines are here, but there is no mechanism for tracking exposure over time.
Administration of exposure will depend on the test setting, and is left to the user.

ONLY WORKS WITH SHADOW TESTING!
}

