% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{createTestBank}
\alias{createTestBank}
\title{Create a test Itembank}
\usage{
createTestBank(model, K = 50, Q = 1, M = 4, between = FALSE)
}
\arguments{
\item{model}{}

\item{K}{}

\item{Q}{}

\item{M}{}
}
\value{
ShadowCAT.itembank
}
\description{
Quick and simple itembanks for testing purposes.
}

