% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Latent trait estimation}
\usage{
estimate(person, test, ...)
}
\arguments{
\item{person}{Person object, see \code{\link{initPerson}}.}

\item{test}{Test object, see \code{\link{initTest}}.}

\item{...}{Additional parameters passed on to the underlying estimation functions. 1}
}
\value{
person Person object, amended with the new estimate.
}
\description{
ML, MAP and EAP estimates.
}
\details{
Obtains a latent trait estimate and variance of the estimate.
}
\section{ML and MAP}{

Maximum Likelihood and Maximum A-Posteriori estimates are based on a Newton-type non-linear minimization algorithm,
and handled with package \code{\link{nlm}}.
}

\section{EAP}{

Estimated A-Posteriori estimates require the repeated evaluation of Q nested integrals, where Q is the dimensionality of the test.
This is performed with adaptive multidimensional Gauss-Hermite quadrature, and handled by package \code{\link{MultiGHQuad}}, see the documentation there for further details.
Note that the number of quadrature points used rises exponentially with the dimensionality of the test - use of EAP estimates with
a 3+ dimensional test may not be a good idea.
}

\section{WML}{

TODO: UPDATE WITH REFERENCES - MORE PRECISE DETAILS.
Note that WML estimation is not included. There is no satisfying solution to multidimensional Weighted Maximum Likelihood Estimation,
current WML estimators as used in other sources do not account for the covariance between dimensions.
}

\section{Variance}{

Variance of the estimate is added to the estimate as an attribute.
}
\examples{
# create a basic test + person
items <- createTestBank("GPCM")
test <- initTest(items, estimator = "ML")
person <- initPerson(items)

# answer a few items
person <- answer(person, test, sample(items$K, 10))

# obtain estimates
ML <- estimate(person, test)$estimate
test$estimator <- "MAP"
MAP <- estimate(person, test)$estimate
test$estimator <- "EAP"
EAP <- estimate(person, test)$estimate
ML; MAP; EAP

# access variance
attr(ML, "variance")

# Note that EAP takes considerably more time when dimensionality is higher...
items5d <- createTestBank("GPCM", Q=5)
test5dEAP <- initTest(items, estimator = "EAP")
test5dMAP <- initTest(items, estimator = "MAP")
person5d <- answer(initPerson(items), test, sample(items5d$K, 10))

system.time(estimate(person5d, test5dEAP))
system.time(estimate(person5d, test5dMAP))
}

