% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_latent_trait.R
\name{estimate_latent_trait}
\alias{estimate_latent_trait}
\title{Latent trait estimation}
\usage{
estimate_latent_trait(estimate, responses, prior, model, administered,
  number_dimensions, estimator, alpha, beta, guessing,
  number_itemsteps_per_item, lower_bound, upper_bound,
  prior_var_safe_nlm = NULL)
}
\arguments{
\item{estimate}{theta estimate with 'variance' as an attribute}

\item{responses}{person responses}

\item{prior}{prior covariance matrix for theta}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively}

\item{administered}{indeces of administered items}

\item{number_dimensions}{number of dimensions}

\item{estimator}{type of estimator to be used, one of "MAP" (Maximum a posteriori estimation), "ML" (maximum likelihood), or "EAP" (Expected A Posteriori Estimation)}

\item{alpha}{matrix of alpha paramteres}

\item{beta}{matrix of beta paramteres}

\item{guessing}{matrix of guessing parameters}

\item{number_itemsteps_per_item}{vector containing the number of non missing cells per row of the beta matrix}

\item{lower_bound}{vector with lower bounds for theta per dimension; estimated theta values smaller than the lowerbound values are truncated to the lowerbound values}

\item{upper_bound}{vector with upper bounds for theta per dimension; estimated theta values larger than the upperbound values are truncated to the upperbound values}

\item{prior_var_safe_nlm}{if not NULL, EAP estimate with prior variance equal to prior_var_safe_ml is computed instead of ML/MAP, if ML/MAP estimate fails}
}
\value{
updated estimate with variance as attribute
}
\description{
ML, MAP and EAP estimates.
}
\details{
Obtains a latent trait estimate and variance of the estimate.
}
\section{ML and MAP}{

Maximum Likelihood and Maximum A-Posteriori estimates are based on a Newton-type non-linear minimization algorithm,
and handled with package \code{\link{nlm}}.
}

\section{EAP}{

Expected A-Posteriori estimates require the repeated evaluation of Q nested integrals, where Q is the dimensionality of the test.
This is performed with adaptive multidimensional Gauss-Hermite quadrature, and handled by package MultiGHQuad, see the documentation there for further details.
Note that the number of quadrature points used rises exponentially with the dimensionality of the test - use of EAP estimates with 
a 3+ dimensional test may not be a good idea.
}

\section{WML}{

TODO: UPDATE WITH REFERENCES - MORE PRECISE DETAILS.
Note that WML estimation is not included. There is no satisfying solution to multidimensional Weighted Maximum Likelihood Estimation,
current WML estimators as used in other sources do not account for the covariance between dimensions.
}

\section{Variance}{

Variance of the estimate is added to the estimate as an attribute.
}
\examples{
number_dimensions <- 1
estimate <- rep(.3, number_dimensions)
model <- "3PLM"
number_items <- 50
responses <- rep(c(1, 0), 17)
administered <- c(6:20, 31:49)
alpha <- matrix(runif(number_items * number_dimensions, .3, 1.5), nrow = number_items, ncol = number_dimensions)
beta <- matrix(rnorm(number_items), nrow = number_items, ncol = 1)
guessing <- c(rep(.1, number_items / 2), rep(.2, number_items / 2))
number_itemsteps_per_item <- number_non_missing_cells_per_row(beta)
lower_bound <- rep(-3, number_dimensions)
upper_bound <- rep(3, number_dimensions)
prior <- diag(1)
prior_var_safe_nlm <- diag(number_dimensions)

# obtain estimates
estimator <- "ML"
ML <- estimate_latent_trait(estimate, responses, prior, model, administered, number_dimensions, estimator, alpha, beta, guessing, number_itemsteps_per_item, lower_bound, upper_bound, prior_var_safe_nlm)
estimator <- "MAP"
MAP <- estimate_latent_trait(estimate, responses, prior, model, administered, number_dimensions, estimator, alpha, beta, guessing, number_itemsteps_per_item, lower_bound, upper_bound, prior_var_safe_nlm)
estimator <- "EAP"
EAP <- estimate_latent_trait(estimate, responses, prior, model, administered, number_dimensions, estimator, alpha, beta, guessing, number_itemsteps_per_item, lower_bound, upper_bound, prior_var_safe_nlm)
ML; MAP; EAP

# access variance
attr(ML, "variance")

# Note that EAP takes considerably more time when dimensionality is higher...
number_dimensions <- 5
estimate <- rep(.3, number_dimensions)
alpha <- matrix(runif(number_items * number_dimensions, .3, 1.5), nrow = number_items, ncol = number_dimensions)
lower_bound <- rep(-3, number_dimensions)
upper_bound <- rep(3, number_dimensions)
prior <- diag(number_dimensions) 

estimator <- "MAP"
system.time(estimate_latent_trait(estimate, responses, prior, model, administered, number_dimensions, estimator, alpha, beta, guessing, number_itemsteps_per_item, lower_bound, upper_bound, prior_var_safe_nlm = diag(number_dimensions)))
estimator <- "EAP"
system.time(estimate_latent_trait(estimate, responses, prior, model, administered, number_dimensions, estimator, alpha, beta, guessing, number_itemsteps_per_item, lower_bound, upper_bound))

}

