% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/info.R, R/info_FI.R
\name{FI}
\alias{FI}
\title{Fisher Information}
\usage{
FI(test, person)

FI(test, person)
}
\arguments{
\item{person}{}

\item{items}{}

\item{items}{}

\item{person}{}
}
\value{
array with an information matrix for each item (QxQxK).

array with an information matrix for each item (QxQxK).
}
\description{
Fisher Information (expected information) per item.

Fisher Information (expected information) per item.
}
\details{
The expected information is the sum of second derivatives, weighted by their probability.

The expected information is the sum of second derivatives, weighted by their probability.
}

