% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowcat_roqua.R
\name{get_beta}
\alias{get_beta}
\title{get beta matrix from beta or eta}
\usage{
get_beta(model, beta, eta)
}
\arguments{
\item{model}{String, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{beta}{matrix containing beta parameters, may be NULL if model is GPCM and eta is defined}

\item{eta}{matrix containing eta parameters, may be NULL if beta is defined}
}
\value{
beta matrix obtained from beta or eta
}
\description{
get beta matrix from beta or eta
}

