% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowcat_roqua.R
\name{get_guessing}
\alias{get_guessing}
\title{get guessing matrix}
\usage{
get_guessing(guessing, beta)
}
\arguments{
\item{guessing}{vector containing guessing parameters; may be NULL in case of zero guessing parameters}

\item{beta}{matrix containing beta parameters}
}
\value{
matrix containing guessing parameters
}
\description{
get guessing matrix
}

