% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_index_max_information.R
\name{get_item_index_max_information}
\alias{get_item_index_max_information}
\title{Maximum Information item selection}
\usage{
get_item_index_max_information(available, item_information)
}
\arguments{
\item{available}{vector with indeces of available items}

\item{item_information}{vector with information of each yet available item, with zeros for administered items (returned by get_item_information() with pad = TRUE)}
}
\value{
item index of item with maximum information.
}
\description{
Naive item selection based on maximum information only.
}
\details{
Selects the item with the highest information.
}

