% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_item.R
\name{get_next_item}
\alias{get_next_item}
\title{Next item}
\usage{
get_next_item(start_items, information_summary, lp_constraints, lp_characters,
  estimate, model, responses, prior, available, administered, number_items,
  number_dimensions, estimator, alpha, beta, guessing,
  number_itemsteps_per_item, lower_bound, upper_bound)
}
\arguments{
\item{start_items}{items that are shown to the patient before adaptive proces starts; one of
list(type = 'random', n)
list(type = 'fixed', indices, n)
list(type = 'random_by_dimension', n_by_dimension, n)
where n = total number of initial items, indices = vector of initial item indeces, 
n_by_dimension = scalar of number of initial items per dimension, or vector with number of initial items for each dimension}

\item{information_summary}{called "objective" by Kroeze; how to summarize information; one of
"D" = determinant: compute determinant(info_sofar_QxQ + info_QxQ_k) for each yet available item k
"PD" = posterior determinant: compute determinant(info_sofar_QxQ_plus_prior + info_QxQ_k) for each yet available item k
"A" = trace: compute trace((info_sofar_QxQ + info_QxQ_k) for each yet available item k
"PA" = posterior trace: compute trace(info_sofar_QxQ_plus_prior + info_QxQ_k) for each yet available item k
"PEKL" = compute Posterior expected Kullback-Leibler Information}

\item{lp_constraints}{data frame with constraints in lp format: the lp_constraints from the list returned by constraints_lp_format(); NULL means no constraints}

\item{lp_characters}{data frame with constraint characters in lp format: the lp_chars from the list returned by constraints_lp_format(); NULL means no constraints}

\item{estimate}{current theta estimate}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{responses}{vector with person responses}

\item{prior}{prior covariance matrix for theta; only required if estimator is "MAP" or "EAP" and output is "likelihoods" or "both"}

\item{available}{vector with indeces of yet available items}

\item{administered}{vector with indeces of administered items}

\item{number_items}{number of items}

\item{number_dimensions}{number of dimensions}

\item{estimator}{type of estimator to be used, one of "MAP" (Maximum a posteriori estimation), "EAP" (Expected A Posteriori Estimation), or "ML" (maximum likelihood)}

\item{alpha}{matrix containing the alpha parameters}

\item{beta}{matrix containing the beta parameters}

\item{guessing}{matrix containing the quessing}

\item{number_itemsteps_per_item}{number_itemsteps_per_item vector containing the number of non missing cells per row of the beta matrix}

\item{lower_bound}{vector with lower bounds for theta per dimension; estimated theta values smaller than the lowerbound values are truncated to the lowerbound values}

\item{upper_bound}{vector with upper bounds for theta per dimension; estimated theta values larger than the upperbound values are truncated to the upperbound values}
}
\value{
integer item index next item
}
\description{
Control function to select the next item, takes account of starting conditions.
}

