% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_expected_kl_information.R
\name{get_posterior_expected_kl_information}
\alias{get_posterior_expected_kl_information}
\title{Posterior expected Kullback-Leibler Information}
\usage{
get_posterior_expected_kl_information(estimate, model, answers, administered,
  available, number_dimensions, estimator, alpha, beta, guessing, prior_form,
  prior_parameters, number_itemsteps_per_item,
  eap_estimation_procedure = "riemannsum")
}
\arguments{
\item{estimate}{Vector containing current theta estimate, with covariance matrix as an attribute.}

\item{model}{One of \code{"3PLM"}, \code{"GPCM"}, \code{"SM"} or \code{"GRM"}, for the three-parameter logistic, generalized partial credit, sequential or graded response model, respectively.}

\item{answers}{Vector with answers to administered items.}

\item{administered}{vector with indices of administered items.}

\item{available}{Vector with indices of available items.}

\item{number_dimensions}{Number of dimensions of theta.}

\item{estimator}{Type of estimator to be used, one of \code{"maximum_likelihood"}, \code{"maximum_aposteriori"}, or \code{"expected_aposteriori"}; see \code{details}.}

\item{alpha}{Matrix of alpha parameters, one column per dimension, one row per item. Row names should contain the item keys. 
Note that so called within-dimensional models still use an alpha matrix, they simply have only one non-zero loading per item.}

\item{beta}{Matrix of beta parameters, one column per item step, one row per item. Row names should contain the item keys. 
Note that \code{shadowcat} expects answer categories to be sequential, and without gaps. That is, the weight parameter in the GPCM model is assumed to be sequential, 
and equal to the position of the 'location' of the beta parameter in the beta matrix.
The matrix should have a number of columns equal to the largest number of item steps over items, items with fewer answer categories should be 
right-padded with \code{NA}. \code{NA} values between answer categories are not allowed, and will lead to errors.
Beta matrix can be set to \code{NULL} if model is GPCM and eta is defined.}

\item{guessing}{Matrix with one column of guessing parameters per item. Row names should contain the item keys. Optionally used in 3PLM model, ignored for all others.}

\item{prior_form}{String indicating the form of the prior; one of \code{"normal"} or \code{"uniform"}.}

\item{prior_parameters}{List containing mu and Sigma of the normal prior: \code{list(mu = ..., Sigma = ...)}, or 
the upper and lower bound of the uniform prior: \code{list(lower_bound = ..., upper_bound = ...)}.
The list element \code{Sigma} should always be in matrix form. List elements \code{mu}, \code{lower_bound}, and \code{upper_bound} should always be vectors.
The length of \code{mu}, \code{lower_bound}, and \code{upper_bound} should be equal to the number of dimensions.
For uniform prior in combination with expected aposteriori estimation, true theta should fall within 
\code{lower_bound} and \code{upper_bound} and be not too close to one of these bounds, in order to prevent errors.}

\item{number_itemsteps_per_item}{Vector containing the number of non missing cells per row of the beta matrix.}

\item{eap_estimation_procedure}{String indicating the estimation procedure if estimator is expected aposteriori and prior form is normal. One of \code{"riemannsum"} for integration via Riemannsum or
\code{"gauss_hermite_quad"} for integration via Gaussian Hermite Quadrature. If prior form is uniform, estimation procedure should always be \code{"riemannsum"}.}
}
\value{
Vector with PEKL information for each yet available item.
Kullback Leibler Divergence for given items and pairs of thetas x posterior density.
returns vector containing information for each yet available item
}
\description{
Kullback-Leibler divergence based on the EAP and ML estimates of ability under the posterior distribution of theta.
Computes the numerical integral of the expectation of KL under the posterior distribution.
}
\details{
Note that even with a simplified grid, the number of quadrature points which have to be calculated for each available item, 
at each step in the CAT is taken to the power Q. Use of KL information is likely to be slow in 3+ dimensional tests.
}

