% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probabilities.R
\name{get_probabilities}
\alias{get_probabilities}
\title{Get matrix with for each included item the probability of scoring in each answer category, given theta}
\usage{
get_probabilities(theta, model, items_to_include, alpha, beta, guessing)
}
\arguments{
\item{theta}{vector with true or estimated theta}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{items_to_include}{vector with indeces of items to include}

\item{alpha}{matrix containing the alpha parameters (for complete test bank)}

\item{beta}{matrix containing the beta parameters (for complete test bank)}

\item{guessing}{matrix containing the quessing parameters (for complete test bank)}
}
\value{
matrix with for each included item (rows) the probability of scoring in each answer category (columns), given theta
}
\description{
Get matrix with for each included item the probability of scoring in each answer category, given theta
}

