% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{not in}
\usage{
x \%not_in\% y
}
\arguments{
\item{x}{vector or value for which it is too be checked if its elements are in vector y}

\item{y}{vector for which it is too be checked if it contains the values in x}
}
\value{
a vector with elements equal to TRUE for each element of x that is not in y
}
\description{
not in
}
\examples{
expect_equal <- c("a", "b", "g", "k", "b") \%not_in\% c("a", "b", "c") == c(FALSE, FALSE, TRUE, TRUE, FALSE);
sum(expect_equal) == 5 || stop("wrong")
}

