% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/selection_MI.R
\name{information}
\alias{information}
\title{Obtain a vector of values of the objective function.}
\usage{
information(test, person)
}
\arguments{
\item{test}{}

\item{person}{}
}
\value{
vector
}
\description{
Obtain a vector of values of the objective function.
}

