% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_latent_trait.R
\name{init_quad}
\alias{init_quad}
\title{Q-dimensional grid of quadrature points}
\usage{
init_quad(Q, prior = list(mu = rep(0, Q), Sigma = diag(Q)), adapt = NULL,
  ip = 6)
}
\arguments{
\item{Q}{Number of dimensions}

\item{prior}{List of prior mean and covariance matrix}

\item{adapt}{List of adaptive mean and covariance matrix; if NULL no adaptation is used}

\item{ip}{Number of quadrature points per dimension. Defaults to 6. Note that the total number of quadrature points is ip^Q}
}
\value{
A list with a matrix X of ip^Q by Q quadrature points and a vector W of length ip^Q associated weights
}
\description{
Creates a flattened, rotated grid that incorporates correlation through an eigenvalue decomposition of the covariance matrix.
Copied from MultiGHQuad package and bugs fixed. Should be moved back to MultiGHQuad package
}
\examples{
grid_points_and_weights1 <- init_quad(Q = 1, prior = list(mu = rep(0, 1), Sigma = diag(1)*2), adapt = list(mu = rep(1, 1), Sigma = diag(1)*5), ip = 50)
all(round(grid_points_and_weights1$X[c(2, 8, 50),], 3) == c(-25.951, -17.382,  30.037)) || stop("wrong")
all(round(grid_points_and_weights1$W[c(2, 8, 50)], 3) == c(-169.013, -76.610, -225.947)) || stop("wrong")
grid_points_and_weights2 <- init_quad(Q = 2, prior = list(mu = rep(0, 2), Sigma = diag(2)*2), adapt = list(mu = rep(1, 2), Sigma = (matrix(rep(2, 4), ncol = 2) + diag(2)*5)), ip = 20)
all(round(grid_points_and_weights2$X[c(2, 8, 50),2], 3) == c(-24.858, -14.749 , -8.557)) || stop("wrong")
all(round(grid_points_and_weights2$W[c(2, 8, 50)], 3) == c(-155.016, -76.948, -40.058)) || stop("wrong")
}

