% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itembank.R, R/itembankClass.R
\name{initItembank}
\alias{initItembank}
\title{initItembank}
\usage{
initItembank(model = "3PLM", alpha = NULL, beta = NULL, guessing = NULL,
  eta = NULL, silent = FALSE)

initItembank(model = "3PLM", alpha = NULL, beta = NULL, guessing = NULL,
  eta = NULL, silent = FALSE)
}
\arguments{
\item{model}{}

\item{alpha}{}

\item{beta}{}

\item{guessing}{}

\item{eta}{}

\item{m}{}

\item{model}{}

\item{alpha}{}

\item{beta}{}

\item{guessing}{}

\item{eta}{}

\item{m}{}
}
\value{
ShadowCAT.items

ShadowCAT.items
}
\description{
Produce an itembank object for a list of parameters

Produce an itembank object for a list of parameters
}

