% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_person.R
\name{initPerson}
\alias{initPerson}
\title{initiate person object for use in ShadowCAT}
\usage{
initPerson(items, theta = rep(0, items$Q), prior = diag(items$Q))
}
\arguments{
\item{items}{Itembank (required for bookkeeping reasons)}

\item{theta}{True theta; only useful for simulations?}

\item{prior}{Prior covariance matrix}
}
\value{
person ShadowCAT.person object
}
\description{
Sets bookkeeping variables and collects person information
}

