% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/info_KL.R
\name{KLB}
\alias{KLB}
\title{Kullback Leibler Divergence for given items and pairs of thetas x posterior density.}
\usage{
KLB(theta, theta0, test, person)
}
\arguments{
\item{theta}{}

\item{theta0}{}

\item{test}{}

\item{person}{}

\item{items}{}
}
\value{
vector info
}
\description{
Kullback Leibler Divergence for given items and pairs of thetas x posterior density.
}

