% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ll_in_nlm_format.R
\name{LL}
\alias{LL}
\title{Log Likelihood}
\usage{
LL(theta, test, person, minimize = FALSE, return_log = TRUE)
}
\arguments{
\item{return_log}{should log of output be returned}

\item{should}{values be reversed (useful for minimization, reverses LL as well as derivatives)}
}
\value{
Log-Likelihood, as well as gradient and hessian attributes.
}
\description{
Internal - fetch appropriate elements from prob for nlm optimizer
}

