% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{matrix_apply}
\alias{matrix_apply}
\title{Matrix apply}
\usage{
matrix_apply(x, margin, FUN, ...)
}
\arguments{
\item{x}{Vector or matrix; if vector, the vector will be converted to a matrix.}

\item{margin}{Vector giving the subscripts the function will be applied over, see \code{\link{apply}} for details.}

\item{FUN}{The funcion to be applied, see \code{\link{apply}} for details.}

\item{...}{Any additional arguments to FUN.}
}
\value{
Output of apply, but always in matrix format.
}
\description{
Apply with input and output converted to matrix
}
\examples{
ShadowCAT:::matrix_apply(c(1:5, NA), 2, sum, na.rm = TRUE)
\dontshow{
ShadowCAT:::matrix_apply(c(1:5, NA), 2, sum, na.rm = TRUE) == matrix(15) || stop("wrong")
}
}

