% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_index_max_information.R
\name{MI}
\alias{MI}
\title{Maximum Information item selection}
\usage{
MI(test, person, item_information)
}
\arguments{
\item{test}{test object, argument is not used and can be removed}

\item{person}{person object}

\item{item_information}{vector with information of each yet available item, with zeros for administered items (returned by objective() with pad = TRUE)}
}
\value{
item index of item with the highest value of the objective function.
}
\description{
Naive item selection based on maximum information only.
}
\details{
Selects the item with the highest value of a specified objective function.
}

