% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_item.R
\name{next_item}
\alias{next_item}
\title{Next item}
\usage{
next_item(person, test)
}
\arguments{
\item{person}{person object}

\item{test}{test object}
}
\value{
integer item index
}
\description{
Control function to select the next item, takes account of starting / stopping conditions.
}

