% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{no_missing_information}
\alias{no_missing_information}
\title{Check both \code{NULL} or not \code{NULL}}
\usage{
no_missing_information(characteristics, constraints)
}
\arguments{
\item{characteristics}{Data frame with characteristics.}

\item{constraints}{List of constraints.}
}
\value{
\code{TRUE} if \code{characteristics} and \code{constraints} are either both \code{NULL} or both not \code{NULL}, \code{FALSE} otherwise.
}
\description{
Check whether characteristics and constraints are either both \code{NULL} or both not \code{NULL}.
}

