% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{number_non_missing_cells_per_row}
\alias{number_non_missing_cells_per_row}
\title{get the number of non-NA cells per row of a matrix}
\usage{
number_non_missing_cells_per_row(X)
}
\arguments{
\item{X}{a matrix}
}
\value{
number of non-NA cells per row
}
\description{
get the number of non-NA cells per row of a matrix
}
\examples{
number_non_missing_cells_per_row(matrix(c(1,2,NA,NA,2,3,NA,2,3), ncol = 3)) == c(1, 3, 2) || stop("wrong")
}

