% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_information.R
\name{objective}
\alias{objective}
\title{Obtain a vector with information for each available item (values of the objective function).}
\usage{
objective(test, person, pad = TRUE)
}
\arguments{
\item{test}{test object}

\item{person}{person object}

\item{pad}{Should the return vector be padded with zeros for items that have already been administered?}
}
\value{
vector
}
\description{
Obtain a vector with information for each available item (values of the objective function).
}

