% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/info_KL.R
\name{PEKL}
\alias{PEKL}
\title{Posterior expected Kullback-Leibler Information}
\usage{
PEKL(test, person, theta_range = -3:3)
}
\arguments{
\item{test}{}

\item{person}{}

\item{simple}{}
}
\value{
vector info
}
\description{
Kullback-Leibler divergence based on the EAP and ML estimates of ability under the posterior distribution of theta.
}

