% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{plotDeriv}
\alias{plotDeriv}
\title{Plot derivatives}
\usage{
plotDeriv(models = c("3PLM", "GRM", "SM", "GPCM"), d = 1,
  estimator = "ML", prior = NULL, theta = seq(-6, 6, length.out = 100),
  K = 10, M = 4)
}
\arguments{
\item{models}{}

\item{d}{}

\item{theta}{}

\item{K}{}

\item{M}{}

\item{estiamtor}{}
}
\description{
Plot derivatives
}

