% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities_and_likelihood.R
\name{probabilities_and_likelihood}
\alias{probabilities_and_likelihood}
\title{Probabilities, likelihood, and derivatives for a given (subset of) items.}
\usage{
probabilities_and_likelihood(theta, responses = NULL, model, items_to_include,
  number_dimensions, estimator, alpha, beta, guessing, prior = NULL,
  return_log_likelihood = TRUE, inverse_likelihood = FALSE,
  output = "probs")
}
\arguments{
\item{theta}{vector with true or estimated theta}

\item{responses}{vector with person responses to the administered items; only required if output is "likelihood" or "both"}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{items_to_include}{vector with indeces of items to include in computations, usually indeces of either the administered items or all items}

\item{number_dimensions}{number of dimensions}

\item{estimator}{type of estimator to be used, one of "maximum_aposteriori", "maximum_likelihood", or "expected_aposteriori"}

\item{alpha}{matrix containing the alpha parameters}

\item{beta}{matrix containing the beta parameters}

\item{guessing}{matrix containing the quessing parameters}

\item{prior}{prior covariance matrix for theta; only required if estimator is "maximum_aposteriori" or "expected_aposteriori" and output is "likelihood" or "both"}

\item{return_log_likelihood}{if TRUE, log of likelihood is returned, else likelihood}

\item{inverse_likelihood}{should likelihood values be reversed (useful for minimization, reverses LL as well as derivatives)}

\item{output}{string, one of "probs" (return matrix with probabilities for each included item), "likelihood" (return likelihood or posterior density of theta, with first and second derivatives as attributes),
or "both" (return list containing both)}
}
\value{
if output = "probs": matrix with for each included item (rows) the probability of scoring in each answer category (columns), given theta
if output = "likelihood": the likelihood (estimator = maximum_likelihood) or posterior density (maximum_aposteriori/expected_aposteriori) of theta, with first and second derivatives as attributes
if output = "both": al list containing both
}
\description{
Probabilities, likelihood, and derivatives for a given (subset of) items.
}

