% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities_and_likelihoods.R
\name{probabilities_and_likelihoods}
\alias{probabilities_and_likelihoods}
\title{Probability and Derivatives for a given (subset of) ShadowCAT.items itembank. It gives the probability of 
responses given theta (only useful for simulation I think), and the likelihood of given responses}
\usage{
probabilities_and_likelihoods(theta, responses = NULL, model,
  items_to_include, number_dimensions, estimator, alpha, beta, guessing,
  prior = NULL, return_log_likelihoods = TRUE,
  inverse_likelihoods = FALSE, output = "probs")
}
\arguments{
\item{theta}{true or estimated theta}

\item{responses}{person responses to the administered items; only required if output is "likelihoods" or "both"}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{items_to_include}{indeces of items to include in computations, usually indeces of either the administered items or all items}

\item{number_dimensions}{number of dimensions}

\item{estimator}{type of estimator to be used, one of "MAP" (Maximum a posteriori estimation), "EAP" (Expected A Posteriori Estimation), or "ML" (maximum likelihood)}

\item{alpha}{matrix containing the alpha parameters}

\item{beta}{matrix containing the beta parameters}

\item{guessing}{matrix containing the quessing}

\item{prior}{prior covariance matrix for theta; only required if estimator is "MAP" or "EAP" and output is "likelihoods" or "both"}

\item{return_log_likelihoods}{if TRUE, log of likelihoods are returned, else likelihoods}

\item{inverse_likelihoods}{should likelihood values be reversed (useful for minimization, reverses LL as well as derivatives)}

\item{output}{string, one of "probs" (return vector of probabilities only), "likelihoods" (return vector of likelihoods with first and second derivatives as attributes),
or "both" (return list of both probabilities and likelihoods with first and second derivatives as attributes)}
}
\value{
probabilities and/or likelihoods, depending on output parameter
}
\description{
Generic
}
\details{
Details
}

