% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{row_or_vector_sums}
\alias{row_or_vector_sums}
\title{Row sum or vector sum}
\usage{
row_or_vector_sums(x)
}
\arguments{
\item{x}{Matrix of which row sums are to be obtained, or vector of which sum is to be obtained.}
}
\value{
Row sums or sum of x.
}
\description{
Get row sums of matrix or sum of vector.
}
\examples{
ShadowCAT:::row_or_vector_sums(matrix(c(1:10), nrow = 2))
ShadowCAT:::row_or_vector_sums(1:5)

\dontshow{
all(ShadowCAT:::row_or_vector_sums(matrix(c(1:10), nrow = 2)) == c(25, 30)) || stop("wrong");
ShadowCAT:::row_or_vector_sums(1:5) == 15 || stop("wrong")
}
}

