% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{rsapply}
\alias{rsapply}
\title{Recursive list apply}
\usage{
rsapply(..., fn)
}
\arguments{
\item{...}{Vectors to loop over.}

\item{fn}{Function to execute for each leaf.}
}
\value{
Lists of lists with the elements as keys and values determined by fn.
}
\description{
Given a bunch of vectors, creates lists of lists with the elements as keys and values determined by fn.
}
\examples{
ShadowCAT:::rsapply(c('a', 'b'), c('f', 'g'), fn=function(x,y) c(x, y))
#> list(a=list(f=c('a', 'f'), g=c('a', 'g')), b=list(f=c('b', 'f'), g=c('b', 'g')))
}

