% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_index_max_information_constrained.R
\name{Shadow}
\alias{Shadow}
\title{Shadow Test item selection}
\usage{
Shadow(test, person, item_information)
}
\arguments{
\item{test}{test object}

\item{person}{person object}

\item{item_information}{item_information vector with information of each yet available item, with zeros for administered items (returned by objective() with pad = TRUE)}
}
\value{
integer item index of item with maximum information within constraints
}
\description{
Based on the Shadow Testing adaptive test assembly procedure by van der Linden (XXX)
}
\details{
At each step in a CAT, a Shadow Test is created consisting of an optimal complete test meeting all constraints.
The best item that was not already administrated is then selected from this test, and presented to the respondent.
}

