% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShadowCAT.R
\name{ShadowCAT}
\alias{ShadowCAT}
\title{ShadowCAT Karel Kroeze}
\usage{
ShadowCAT(person, test, verbose = FALSE)
}
\arguments{
\item{verbose}{if larger than 0, print estimate and variance of estimate after test is finished. If larger than 1, also print estimate and variance of estimate at each iteration}
}
\value{
person
}
\description{
Run test with a specified person, and a specified test. See initPerson and initTest.
 This is a simple wrapper to call the right methods, options should be defined in the test object.
}
\details{
Details
}

