% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functions.R
\name{simulate_answer}
\alias{simulate_answer}
\title{Simulates answers}
\usage{
simulate_answer(theta, model, alpha, beta, guessing, item_keys)
}
\arguments{
\item{theta}{Vector with true theta.}

\item{model}{One of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{alpha}{Matrix of alpha parameters. See \code{shadowcat} for details.}

\item{beta}{Matrix of beta parameters, one column per item step, one row per item. See \code{shadowcat} for details.}

\item{guessing}{One column matrix of guessing parameters per item. Row names should contain the item keys. Optionally used in 3PLM model, ignored for all others.}

\item{item_keys}{Character vector of item keys for which answers should be simulated.}
}
\value{
Vector with responses
}
\description{
Simulates answers on specified items, given true theta. Only for testing purposes
}
\examples{
item_bank <- simulate_testbank("GPCM")
simulate_answer(theta = .3, model = "GPCM", alpha = item_bank$alpha, beta = item_bank$beta, guessing = NULL, item_keys = "item3")
}

