% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ShadowCAT.R
\name{stop_test}
\alias{stop_test}
\title{Done}
\usage{
stop_test(person, test)
}
\arguments{
\item{person}{}

\item{test}{}
}
\value{
boolean completed
}
\description{
Control function to check if the test is completed.
}

