% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_shadowcat_items.R
\name{subset.ShadowCAT.items}
\alias{subset.ShadowCAT.items}
\title{Subset method for ShadowCAT.items.}
\usage{
\method{subset}{ShadowCAT.items}(x, subset, ...)
}
\arguments{
\item{x}{item characteristics class ShadowCAT.items (as returned by initItembank()).}

\item{subset}{indeces of items to be selected}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
Item subset of provided itembank, with class ShadowCAT.items, as well as any original classes.
}
\description{
Creates a valid subset of items to work with in ShadowCAT functions. Primarily used internally.
}

