% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{sum_loop_outputs}
\alias{sum_loop_outputs}
\title{get sum of objects returned by a loop, added to a starting object}
\usage{
sum_loop_outputs(start_object, loop_vector, FUN, ...)
}
\arguments{
\item{start_object}{the object to which the sum is to be added}

\item{loop_vector}{the vector to be looped over}

\item{FUN}{a function of the loop_vector values, returning the object to be added to start_object at each iteration. First argument should be loop values argument}

\item{...}{any additional arguments to FUN}
}
\description{
get sum of objects returned by a loop, added to a starting object
}
\examples{
X = matrix(1:6, ncol = 2);
sum_loop_outputs(matrix(0, 2, 2), 1:3, FUN = function(item, X) { X[item,] \%*\% t(X[item,]) }, X = X) == (X[1,] \%*\% t(X[1,]) + X[2,] \%*\% t(X[2,]) + X[3,] \%*\% t(X[3,])) || stop("wrong")
}

