% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_must_stop.R
\name{test_must_stop}
\alias{test_must_stop}
\title{Done}
\usage{
test_must_stop(number_responses, variance_current_estimate, max_n,
  stop_variance_target = NULL)
}
\arguments{
\item{number_responses}{number of responses given}

\item{variance_current_estimate}{variance(s) of the current estimate(s)}

\item{max_n}{test length at which testing should stop (even if target has not been reached yet in case of variance stopping rule; n <= 0 never stops for max length)}

\item{stop_variance_target}{a vector with the variance target for each dimension; if null, variance is not taken into account}
}
\value{
boolean completed
}
\description{
Control function to check if the test is completed.
}

