% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functions.R
\name{test_shadowcat}
\alias{test_shadowcat}
\title{Simulate a testing routine with shadowcat}
\usage{
test_shadowcat(true_theta, prior_form, prior_parameters, model, alpha, beta,
  guessing, eta = NULL, start_items, stop_test, estimator,
  information_summary, constraints_and_characts = NULL, lower_bound = NULL,
  upper_bound = NULL, safe_eap = FALSE, initital_estimate = rep(0,
  ncol(alpha)), initial_variance = diag(ncol(alpha)) * 25,
  eap_estimation_procedure = "riemannsum")
}
\arguments{
\item{true_theta}{True theta value or vector.}

\item{prior_form}{String indicating the form of the prior; one of \code{"normal"} or \code{"uniform"}. Not required if estimator is maximum likelihood.}

\item{prior_parameters}{List containing mu and Sigma of the normal prior: \code{list(mu = ..., Sigma = ...)}, or 
the upper and lower bound of the uniform prior: \code{list(lower_bound = ..., upper_bound = ...)}. Not required if estimator is maximum likelihood.
See \code{\link{shadowcat}} for details.}

\item{model}{One of \code{"3PLM"}, \code{"GPCM"}, \code{"SM"} or \code{"GRM"}, for the three-parameter logistic, generalized partial credit, sequential or graded response model, respectively.}

\item{alpha}{Matrix of alpha parameters. See \code{\link{shadowcat}} for details.}

\item{beta}{Matrix of beta parameters. See \code{\link{shadowcat}} for details.}

\item{guessing}{Matrix with one column of guessing parameters per item. See \code{\link{shadowcat}} for details.}

\item{eta}{Matrix of location parameters, optionally used in GPCM model, ignored for all others. See \code{\link{shadowcat}} for details.}

\item{start_items}{List indicating the items that should be shown to the respondent before the theta estimate will be updated
for the first time. See \code{\link{shadowcat}} for details.}

\item{stop_test}{List indicating rules for when to terminate the test. See \code{\link{shadowcat}} for details.}

\item{estimator}{Type of estimator to be used, one of \code{"maximum_likelihood"}, \code{"maximum_aposteriori"}, or \code{"expected_aposteriori"}.}

\item{information_summary}{How to summarize Fisher information, used for selection of item with maximum information. One of
\code{"determinant"}, \code{"posterior_determinant"}, \code{"trace"}, \code{"posterior_trace"}, or \code{"posterior_expected_kullback_leibler"}.}

\item{constraints_and_characts}{List with constraints and characteristics. See \code{\link{shadowcat}} for details.}

\item{lower_bound}{Vector with lower bounds for theta per dimension. See \code{\link{shadowcat}} for details.}

\item{upper_bound}{Vector with upper bounds for theta per dimension. See \code{\link{shadowcat}} for details.}

\item{safe_eap}{Only relevant if estimator is expected aposteriori. 
Set to \code{TRUE} if estimator should switch to maximum aposteriori if the integration algorithm results in an error.
An error may occur if the prior is uniform, estimator is expected aposteriori, and the bounds of the prior do not exceed the true theta value, or are too close to it.}

\item{initital_estimate}{Vector containing the initial theta estimate, before any items have been administered.}

\item{initial_variance}{Matrix containing the initial covariance matrix, before any items have been administered.}

\item{eap_estimation_procedure}{String indicating the estimation procedure if estimator is expected aposteriori and prior form is normal. One of \code{"riemannsum"} 
for integration via Riemannsum or \code{"gauss_hermite_quad"} for integration via Gaussian Hermite Quadrature. 
If prior form is uniform, estimation procedure should always be \code{"riemannsum"}.}
}
\value{
List as returned by \code{\link{shadowcat}} after test is terminated, with \code{variance} element turned into matrix.
}
\description{
Simulate several iterations of getting the key of the new item to administer and the updated theta estimate, and getting an answer to the new item.
}
\examples{
# One dimension
alpha_beta_one_dim <- simulate_testbank(model = "GPCM", number_items = 50, number_dimensions = 1, number_itemsteps = 3)
test_shadowcat(true_theta = 2, prior_form = "normal", prior_parameters = list(mu = 0, Sigma = diag(1)), model = "SM", alpha = alpha_beta_one_dim$alpha, beta = alpha_beta_one_dim$beta, guessing = NULL, start_items = list(type = 'random', n = 3), stop_test = list(max_n = 20, target = 0.1), estimator = "maximum_aposteriori", information_summary = "posterior_determinant")

# Three dimensions
alpha_beta_three_dim <- simulate_testbank(model = "GPCM", number_items = 100, number_dimensions = 3, number_itemsteps = 3)
test_shadowcat(true_theta = c(0, 1, -.5), prior_form = "normal", prior_parameters = list(mu = c(0, 0, 0), Sigma = diag(3)), model = "SM", alpha = alpha_beta_three_dim$alpha, beta = alpha_beta_three_dim$beta, guessing = NULL, start_items = list(type = 'random', n = 3), stop_test = list(max_n = 60, target = c(.1, .1, .1)), estimator = "maximum_aposteriori", information_summary = "posterior_determinant")
}

