% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{testCAT}
\alias{testCAT}
\title{Comprehensive testing wrapper.}
\usage{
testCAT(models = c("3PLM", "GRM", "GPCM", "SM"), estimators = c("ML",
  "MAP"), dims = c(1, 2, 3), between = c(TRUE, FALSE), covars = NULL,
  K = 100, start = list(type = "random", n = 5), stop = list(type =
  "length", n = 30))
}
\arguments{
\item{models}{}

\item{estimators}{}

\item{dims}{}

\item{size}{}

\item{length}{}
}
\description{
Test a CAT run under various conditions.
}

