% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{testDeriv}
\alias{testDeriv}
\title{Test analytical derivatives}
\usage{
testDeriv(models = c("3PLM", "GPCM", "GRM", "SM"), dims = 1:3,
  eval.points = c(-2, 0, 2), K = 10, estimator = "ML")
}
\arguments{
\item{models}{}

\item{dims}{}

\item{eval.points}{}

\item{K}{}

\item{estimator}{}
}
\description{
Compare analytical derivatives with their numerical counterparts
}

