% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{testEstimators}
\alias{testEstimators}
\title{Test Estimators}
\usage{
testEstimators(models = c("3PLM", "GRM", "GPCM", "SM"), estimators = c("ML",
  "MAP", "EAP"), dims = 1:3, between.dim = c(TRUE, FALSE),
  theta = lapply(dims, function(x) rep(0, x)), prior = lapply(dims,
  function(x) diag(x)), N = 100, K = 20, detail = FALSE, silent = FALSE)
}
\arguments{
\item{models}{}

\item{estimators}{}

\item{dims}{}

\item{theta}{}

\item{prior}{}

\item{N}{}

\item{K}{}

\item{detail}{}

\item{between}{}
}
\description{
Test the various estimators
}

