% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{testFI}
\alias{testFI}
\title{Test Information Measures}
\usage{
testFI(models = c("3PLM", "GPCM", "SM", "GRM"), dims = 1:3,
  methods = c("FI", "PFI"), K = 1, show.items = FALSE)
}
\arguments{
\item{models}{}

\item{dims}{}

\item{methods}{}

\item{K}{}
}
\description{
Test Information Measures
}

