% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{testProb}
\alias{testProb}
\title{Test probability functions}
\usage{
testProb(models = c("3PLM", "GPCM", "SM", "GRM"), Q = 1, alpha = 1,
  beta = 0)
}
\arguments{
\item{models}{}

\item{Q}{}

\item{alpha}{}

\item{beta}{}
}
\description{
Test probability functions
}

