% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{transpose_if_ncol_and_nrow_larger_1}
\alias{transpose_if_ncol_and_nrow_larger_1}
\title{transpose matrix if number of columns is larger than 1}
\usage{
transpose_if_ncol_and_nrow_larger_1(X)
}
\arguments{
\item{X}{a matrix}
}
\value{
transpose of X if number of columns and number of rows are larger than 1, X otherwise
}
\description{
transpose matrix if number of columns is larger than 1
}
\examples{
expect_equal <- transpose_if_ncol_and_nrow_larger_1(matrix(1:50, ncol = 2)) == t(matrix(1:50, ncol = 2));
sum(expect_equal) == 50 || stop("wrong")
}

