% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_and_run.R
\name{validate_and_run}
\alias{validate_and_run}
\title{Template to structure and test larger methods}
\usage{
validate_and_run()
}
\description{
First validate is called, if any errors are added, the value of invalid_result is returned, else the value of result.
Loads the following methods into the environment of the calling method:
  add_error(key, value)
  errors() #returns list(key=value, key2=value2)
  validate_and_runner() # Used internally, contains the logic of functions to call.
}
\examples{
two_times_two <- function(two, also_two) {
  result <- function() helper()

  helper <- function() two * also_two

  validate <- function() {
    if(two != 2) add_error('two', 'should be 2')
    if(also_two != 2) add_error('also_two', 'should also be two')
  }

  invalid_result <- function() {
    list(result=NA, errors=errors())
  }

  ShadowCAT:::validate_and_run()
}
two_times_two(2,2) == 4 || stop('not 4')
two_times_two(3,2)$errors$two == 'should be 2' || stop('wrong result')
}

