% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_and_run.R
\name{validate_and_run.test}
\alias{validate_and_run.test}
\title{Get result function}
\usage{
validate_and_run.test(fn, ...)
}
\arguments{
\item{fn}{The function to get the function result of.}

\item{...}{Arguments for fn.}
}
\value{
The result function from the calling method.
}
\description{
Returns the result function instead of the result of the result function.
}
\examples{
fn <- function(a) {
  result <- function() helper()
  helper <- function() 2*a
  ShadowCAT:::validate_and_run()
}
tst <- ShadowCAT:::validate_and_run.test(fn, 4)
helper <- get('helper', environment(tst))
helper() == 8 || stop('error')
}

