% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variance.R
\name{variance}
\alias{variance}
\title{Variance}
\usage{
variance(person, test, ...)
}
\arguments{
\item{person}{}

\item{test}{}

\item{...}{additional internal arguments}
}
\value{
person
}
\description{
Computes the (asymptotic) variance-covariance matrix for the current estimate
}

