% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_answer.R
\name{answer}
\alias{answer}
\title{Simulate response(s)}
\usage{
answer(person, test, indeces)
}
\arguments{
\item{person}{Person object, see \code{\link{initPerson}}.}

\item{test}{Test object, see \code{\link{initPerson}}.}

\item{indeces}{If not NULL, answer questions with the given indeces, and update the person object. If NULL, answers the full test.}
}
\value{
vector responses, or updated person object if indeces is set.
}
\description{
Given a person and (a subset of) an itembank, simulate a response pattern.
}
\examples{
items <- createTestBank("GPCM")
test <- initTest(items)
person <- initPerson(items)

# simulates responses to all questions, and returns a vector response pattern.
answer(person, test, 1:50)

# simulates responses to the specified question indeces, and returns an updated person object.
answer(person, test, sample(test$items$K, 5))


}

