% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_answer.R
\name{answer}
\alias{answer}
\title{Simulates responses on items indicated by indeces, given true theta}
\usage{
answer(theta, model, number_dimensions, estimator, alpha, beta, guessing,
  number_itemsteps, indeces)
}
\arguments{
\item{theta}{true theta}

\item{model}{String, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{number_dimensions}{number of dimensions}

\item{estimator}{type of estimator to be used, one of "MAP" (Maximum a posteriori estimation) or "ML" (maximum likelihood); 
"EAP" (Expected A Posteriori Estimation) is currently not working due to problems with the MultiGHQuad package}

\item{alpha}{matrix of alpha parameters}

\item{beta}{matrix of beta parameters}

\item{guessing}{vector of guessing parameters}

\item{number_itemsteps}{number of itemsteps}

\item{indeces}{answer questions with the given indeces}
}
\value{
vector responses, or updated person object if indeces is set.
}
\description{
Simulates responses on items indicated by indeces, given true theta
}
\examples{
items <- simulate_testbank("GPCM")

# simulates responses on items indicated by indeces, given true theta
answer(.3, "GPCM", 1, "MAP", items$pars$alpha, items$pars$beta, items$pars$guessing, items$M, 3)

}

