% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{as.scalar2}
\alias{as.scalar2}
\title{Return as a scalar
NULL -> NA}
\usage{
as.scalar2(obj)
}
\arguments{
\item{obj}{Any object}
}
\value{
NA (if NULL) or obj as a scalar
}
\description{
Return as a scalar
NULL -> NA
}

