% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{categorical_to_dummy}
\alias{categorical_to_dummy}
\title{transform categorical vector into dummy variables}
\usage{
categorical_to_dummy(unique_values, categorical_vector)
}
\arguments{
\item{unique_values}{a vector containing the unique values of the categorical vector}

\item{categorical_vector}{the categorical vector to be transformed into dummy variables}
}
\value{
a matrix containing the dummy variables, number of columns (dummies) equals the number of unique values
}
\description{
transform categorical vector into dummy variables
}
\examples{
expect_equal <- categorical_to_dummy(c("a", "b", "c"), rep(c("a", "b", "c"), 3))[1,] == rep(c(1, 0, 0), 3);
sum(expect_equal) == 9 || stop("wrong")
}

