% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_lp_format.R
\name{constraints_lp_format}
\alias{constraints_lp_format}
\title{Creates a list with characteristics and constraints which can be used in lp function from lpSolve package.}
\usage{
constraints_lp_format(max_n, number_items, characteristics = NULL,
  constraints = NULL)
}
\arguments{
\item{max_n}{test length at which testing should stop}

\item{number_items}{number of items available in the item bank}

\item{characteristics}{\code{data.frame} with characteristics, see \code{details}.}

\item{constraints}{\code{list} of constraints, see \code{details}.}
}
\value{
list containing characteristics and constraints in lp format; 
the maximum test length is always included as an additional constraint; see \code{details}.
}
\description{
Creates a list with characteristics and constraints which can be used in lp function from lpSolve package.
}
\section{Characteristics specification}{

\code{characteristics} should be specified as a data.frame of characteristics. Each row indicates the characteristics of
one item. Each column indicates how all items score on a certain characteristic.
Characteristics may be categorical or numeric.
}

\section{Constraint specification}{

\code{constraints} should be specified as a list of constraints, each constraint is a list with three named values;
\code{name} the column name of the characteristic this constraint applies to. For categorical characteristics the level should be specified as \code{name/value},
where \code{name} is the column name of the characteristic and \code{value} is the specific level of the characteristic this constraint applies to. 
\code{op} the logical operator to be used. Valid options are "<", "=", ">" and "><".
\code{target} the target value, numeric. It indicates the number of items of the relevant characteristic that should be administered ("="), or
minimally (">"), maximally ("<"), or minimally and maximally ("><"; vector with two values required) administered.
}

\section{Return object}{

The constraints and characteristics in lp format will be stored within a list. Setting the constraints and characteristics
arguments to NULL results in a list with only the test length constraint.
Constraints in Shadow Tests are implemented through linear programming, for which the package lpSolve is used. 
The returned constraints object is a list with two named elements; lp_constraints and lp_chars. The lp_constraints 
and lp_chars are set up to work with lpSolve, and should not be manually edited.
}

\section{Note about length}{

Note that the maximum test length is always included as an additional constraint.
}
\examples{
max_n <- 30
number_items <- 50

# set up some dummy characteristics.
content <- sample(c('algebra','physics','calculus'), number_items, TRUE)
time <- rnorm(number_items)
exclusive <- rep(0, number_items)
exclusive[sample(number_items, 4)] <- 1

# bind them in a data.fame
characteristics <- data.frame(content, time, exclusive)

# set up the constraints
constraints <- list(
  list(name = 'content/algebra',
       op = '><',
       target = c(5,10)),
  list(name = 'content/physics',
       op = '><',
       target = c(2,5)),
  list(name = 'time',
       op = '<',
       target = 20),
  list(name = 'exclusive',
       op = '<',
       target = 2))

# get list of characteristics and constraintrs in lp format
chars_constraints_lp <- constraints_lp_format(max_n, number_items, characteristics, constraints)

}

