% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/constraintsClass.R
\name{createConstraints}
\alias{createConstraints}
\title{Convenience wrapper to create a constraints list for use with ShadowCAT.}
\usage{
createConstraints(characteristics, constraints = NULL)
}
\arguments{
\item{characteristics}{\code{data.frame} with one characteristic per column, one item per row.
Deals with factor, string and numeric characteristics. See Details.}

\item{constraints}{\code{data.frame} with three columns.
\code{characteristic} name of the characteristic in the \code{characteristics} data.frame, for categorical variables each level should have its own seperate entry in the format \code{variable.level}.
\code{operator} Mathematical operator to be used. Can be any of \code{=, <= or >=} (\code{==, < and >} are equivalent to these).
\code{target} The target value of the constraint.}
}
\value{
constrains, \code{data.frame} of constraints with class ShadowCAT.constraints, to be used in adaptive testing.
}
\description{
Convenience wrapper to create a constraints list for use with ShadowCAT.
}

