% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initItembank.R
\name{createTestBank}
\alias{createTestBank}
\title{Create a test Itembank}
\usage{
createTestBank(model, K = 50, Q = 1, M = 4, between = FALSE)
}
\arguments{
\item{model}{String, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{K}{number of items}

\item{Q}{number of dimensions}

\item{M}{number of item steps (number of categories minus 1)}

\item{between}{is TRUE, force items to load on one dimension each.}
}
\value{
ShadowCAT.itembank
}
\description{
Quick and simple itembanks for testing purposes.
}

