% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Latent trait estimation}
\usage{
estimate(person, test, ...)
}
\arguments{
\item{person}{}

\item{test}{}

\item{...}{}
}
\value{
person
}
\description{
ML, MAP, EAP
}

