% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_item.R
\name{get_best_item}
\alias{get_best_item}
\title{Select best item}
\usage{
get_best_item(information_summary, lp_constraints, lp_characters, estimate,
  model, answers, prior_form, prior_parameters, available, administered,
  number_items, number_dimensions, estimator, alpha, beta, guessing,
  number_itemsteps_per_item, stop_test,
  eap_estimation_procedure = "riemannsum")
}
\arguments{
\item{information_summary}{How to summarize Fisher information, used for selection of item with maximum information. One of
\code{"determinant"}, \code{"posterior_determinant"}, \code{"trace"}, \code{"posterior_trace"}, or \code{"posterior_expected_kullback_leibler"}.}

\item{lp_constraints}{Data frame with constraints in \code{\link{lp}} format: the \code{lp_constraints} from the list returned by \code{\link{constraints_lp_format}}. \code{NULL} means no constraints.}

\item{lp_characters}{Data frame with constraint characteristics in \code{\link{lp}} format: the \code{lp_chars} from the list returned by \code{\link{constraints_lp_format}}. \code{NULL} means no constraints.}

\item{estimate}{Vector containing current theta estimate, with covariance matrix as an attribute.}

\item{model}{One of \code{"3PLM"}, \code{"GPCM"}, \code{"SM"} or \code{"GRM"}, for the three-parameter logistic, generalized partial credit, sequential or graded response model, respectively.}

\item{answers}{Vector with answers to administered items.}

\item{prior_form}{String indicating the form of the prior; one of \code{"normal"} or \code{"uniform"}. Not required if estimator is maximum likelihood.}

\item{prior_parameters}{List containing mu and Sigma of the normal prior: \code{list(mu = ..., Sigma = ...)}, or 
the upper and lower bound of the uniform prior: \code{list(lower_bound = ..., upper_bound = ...)}. Not required if estimator is maximum likelihood.
The list element \code{Sigma} should always be in matrix form. List elements \code{mu}, \code{lower_bound}, and \code{upper_bound} should always be vectors.
The length of \code{mu}, \code{lower_bound}, and \code{upper_bound} should be equal to the number of dimensions.
For uniform prior in combination with expected aposteriori estimation, true theta should fall within 
\code{lower_bound} and \code{upper_bound} and be not too close to one of these bounds, in order to prevent errors. 
Setting the function argument \code{safe_eap} to \code{TRUE} ensures that the estimation switches to maximum aposteriori if the expected aposteriori estimate fails.}

\item{available}{Vector with indices of yet available items.}

\item{administered}{Vector with indices of administered items.}

\item{number_items}{Number of items in the test bank.}

\item{number_dimensions}{Number of dimensions.}

\item{estimator}{Type of estimator to be used, one of \code{"maximum_likelihood"}, \code{"maximum_aposteriori"}, or \code{"expected_aposteriori"}; see \code{details}.}

\item{alpha}{Matrix of alpha parameters, one column per dimension, one row per item. Row names should contain the item keys. 
Note that so called within-dimensional models still use an alpha matrix, they simply have only one non-zero loading per item.}

\item{beta}{Matrix of beta parameters, one column per item step, one row per item. Row names should contain the item keys. 
Note that \code{shadowcat} expects answer categories to be sequential, and without gaps. That is, the weight parameter in the GPCM model is assumed to be sequential, 
and equal to the position of the 'location' of the beta parameter in the beta matrix.
The matrix should have a number of columns equal to the largest number of item steps over items, items with fewer answer categories should be 
right-padded with \code{NA}. \code{NA} values between answer categories are not allowed, and will lead to errors.
Beta matrix can be set to \code{NULL} if model is GPCM and eta is defined.}

\item{guessing}{Matrix with one column of guessing parameters per item. Row names should contain the item keys. Optionally used in 3PLM model, ignored for all others.}

\item{number_itemsteps_per_item}{Vector containing the number of non missing cells per row of the beta matrix.}

\item{stop_test}{List indicating rules for when to terminate the test. Should be a list of the form
\code{list(target = ..., max_n = ..., min_n = ..., cutoffs = ...)}, 
where \code{target} is a vector indicating the maximum acceptable variance per dimension; \code{NULL} means no variance target,
\code{max_n} is the test length at which the test should be terminated (even if the target has not been reached yet), 
\code{min_n} is the minimum test length; \code{NULL} means no mimimum test length, and
\code{cutoffs} is a matrix containing cut off values per dimension (columns) and test iteration (rows). First row contains cut off values for when no items have been
administered yet, second row for when one item has been administered, etc. If estimate + 3SE < cutoff for each dimension at a certain iteration, test terminates; 
\code{NULL} means no cut off values.}

\item{eap_estimation_procedure}{String indicating the estimation procedure if estimator is expected aposteriori and prior form is normal. One of \code{"riemannsum"} for integration via Riemannsum or
\code{"gauss_hermite_quad"} for integration via Gaussian Hermite Quadrature. If prior form is uniform, estimation procedure should always be \code{"riemannsum"}.}
}
\value{
Item index of best item.
}
\description{
Select the item with maximum information from the available items, taking constraints into account.
}

