% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_item.R
\name{get_best_item}
\alias{get_best_item}
\title{Next item in an adaptive test. Takes a person and test object, and returns the index of the next item based}
\usage{
get_best_item(item_selection, information_summary, lp_constraints,
  lp_characters, estimate, model, responses, prior, available, administered,
  number_items, number_dimensions, estimator, alpha, beta, guessing,
  number_itemsteps_per_item, lower_bound, upper_bound)
}
\arguments{
\item{item_selection}{selection criterion; one of "MI" (maximum information) or "Shadow" (maximum information and take constraints into account)}

\item{information_summary}{called "objective" by Kroeze; how to summarize information; one of
"D" = determinant: compute determinant(info_sofar_QxQ + info_QxQ_k) for each yet available item k
"PD" = posterior determinant: compute determinant(info_sofar_QxQ_plus_prior + info_QxQ_k) for each yet available item k
"A" = trace: compute trace((info_sofar_QxQ + info_QxQ_k) for each yet available item k
"PA" = posterior trace: compute trace(info_sofar_QxQ_plus_prior + info_QxQ_k) for each yet available item k
"PEKL" = compute Posterior expected Kullback-Leibler Information}

\item{lp_constraints}{data frame with constraints in lp format: the constraints list returned by constraints_correct_format()}

\item{lp_characters}{data frame with constraint characters in lp format: the lp_chars list returned by constraints_correct_format()}

\item{estimate}{current theta estimate}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{responses}{vector with person responses}

\item{prior}{prior covariance matrix for theta; only required if estimator is "MAP" or "EAP" and output is "likelihoods" or "both"}

\item{available}{vector with indeces of yet available items}

\item{administered}{vector with indeces of administered items}

\item{number_items}{number of items}

\item{number_dimensions}{number of dimensions}

\item{estimator}{type of estimator to be used, one of "MAP" (Maximum a posteriori estimation), "EAP" (Expected A Posteriori Estimation), or "ML" (maximum likelihood)}

\item{alpha}{matrix containing the alpha parameters}

\item{beta}{matrix containing the beta parameters}

\item{guessing}{matrix containing the quessing}

\item{number_itemsteps_per_item}{number_itemsteps_per_item vector containing the number of non missing cells per row of the beta matrix}

\item{lower_bound}{vector with lower bounds for theta per dimension; estimated theta values smaller than the lowerbound values are truncated to the lowerbound values}

\item{upper_bound}{vector with upper bounds for theta per dimension; estimated theta values larger than the upperbound values are truncated to the upperbound values}
}
\value{
integer item index
}
\description{
This function is a wrapper that sends the actual work to the correct subroutines.
}

