% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_item.R
\name{get_best_item}
\alias{get_best_item}
\title{Best item in an adaptive test.}
\usage{
get_best_item(information_summary, lp_constraints, lp_characters, estimate,
  model, answers, prior_form, prior_parameters, available, administered,
  number_items, number_dimensions, estimator, alpha, beta, guessing,
  number_itemsteps_per_item, stop_test,
  eap_estimation_procedure = "riemannsum")
}
\arguments{
\item{information_summary}{called "objective" by Kroeze; how to summarize information; one of
"determinant": compute determinant(info_sofar_QxQ + info_QxQ_k) for each yet available item k
"posterior_determinant": compute determinant(info_sofar_QxQ_plus_prior_information + info_QxQ_k) for each yet available item k
"trace": compute trace((info_sofar_QxQ + info_QxQ_k) for each yet available item k
"posterior_trace": compute trace(info_sofar_QxQ_plus_prior_information + info_QxQ_k) for each yet available item k
"posterior_expected_kullback_leibler" = compute Posterior expected Kullback-Leibler Information}

\item{lp_constraints}{data frame with constraints in lp format: the lp_constraints from the list returned by constraints_lp_format(); NULL means no constraints}

\item{lp_characters}{data frame with constraint characters in lp format: the lp_chars from the list returned by constraints_lp_format(); NULL means no constraints}

\item{estimate}{vector containing current theta estimate}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{answers}{vector with person answers}

\item{prior_form}{String indicating the form of the prior; one of "normal" or "uniform"}

\item{prior_parameters}{List containing mu and Sigma of the normal prior: list(mu = ..., Sigma = ...), or 
the upper and lower bound of the uniform prior: list(lower_bound = ..., upper_bound = ...). Sigma should always
be in matrix form.}

\item{available}{vector with indices of yet available items}

\item{administered}{vector with indices of administered items}

\item{number_items}{number of items in the test bank}

\item{number_dimensions}{number of dimensions}

\item{estimator}{type of estimator to be used, one of "maximum_aposteriori", "maximum_likelihood", or "expected_aposteriori"}

\item{alpha}{matrix containing the alpha parameters}

\item{beta}{matrix containing the beta parameters}

\item{guessing}{matrix containing the quessing parameters}

\item{number_itemsteps_per_item}{vector containing the number of non missing cells per row of the beta matrix}

\item{stop_test}{rule for when to stop providing new items to patient; should be a list of the form
list(target = ..., max_n = ..., min_n = ..., cutoffs = ...), 
where max_n = test length at which testing should stop (even if target has not been reached yet in case of variance stopping rule), 
target = vector of maximum acceptable variances per dimension; NULL means no variance target,
min_n = minimum test length; NULL means no mimimum test length,
cutoffs = matrix containing cut off values per dimension (columns) and test iteration (rows). First row contains cut off values for when no items have been
administered yet, second row for when one item has been administered, etc. If estimate + 3SE < cutoff for each dimension at certain iteration, test stops; 
NULL means no cut off values}

\item{eap_estimation_procedure}{String indicating the estimation procedure for the expected aposteriori estimate, which is computed
in get_posterior_expected_kl_information() if it is not the requested estimator in shadowcat(). One of "riemannsum" for integration via Riemannsum or
"gauss_hermite_quad" for integration via Gaussian Hermite Quadrature. Only important here if information_summary is posterior_expected_kl_information.}
}
\value{
integer item index of best item
}
\description{
This function is a wrapper that sends the actual work to the correct subroutines.
}

