% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eap_estimate_riemannsum.R
\name{get_eap_estimate_riemannsum}
\alias{get_eap_estimate_riemannsum}
\title{Compute the expected aposteriori estimate of the latent trait theta, with the posterior covariance matrix as an attribute}
\usage{
get_eap_estimate_riemannsum(dimension, likelihood, prior_form, prior_parameters,
  adapt = NULL, number_gridpoints = 50, ...)
}
\arguments{
\item{dimension}{Number of dimensions}

\item{likelihood}{Likelihood function, first argument should be theta}

\item{prior_form}{String indicating the form of the prior; one of "normal" or "uniform"}

\item{prior_parameters}{List containing mu and Sigma of the normal prior: list(mu = ..., Sigma = ...), or 
the upper and lower bound of the uniform prior: list(lower_bound = ..., upper_bound = ...). Sigma should always
be in matrix form.}

\item{adapt}{List containing mu and Sigma for the adaptation of the grid points: list(mu = ..., Sigma = ...).
If NULL, adaptation with normal prior is based on the prior parameters, and no adaptation is made with uniform prior.}

\item{number_gridpoints}{Value indicating the number of grid points to use for the Riemannsum}

\item{...}{Any additional arguments to likelihood}
}
\description{
Integration approximation occurs via a Riemannsumm, where grid points can be adapted to the location of the
posterior distribution
}

