% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fisher_information.R
\name{get_fisher_information}
\alias{get_fisher_information}
\title{Fisher Information}
\usage{
get_fisher_information(estimate, model, number_dimensions, alpha, beta,
  guessing, number_itemsteps_per_item)
}
\arguments{
\item{estimate}{vector containing theta estimate}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{number_dimensions}{number of dimensions}

\item{alpha}{matrix containing the alpha parameters}

\item{beta}{matrix containing the beta parameters}

\item{guessing}{matrix containing the quessing parameters}

\item{number_itemsteps_per_item}{vector containing the number of non missing cells per row of the beta matrix}
}
\value{
array with an information matrix for each item (QxQxK).
}
\description{
Fisher Information (expected information) per item.
}
\details{
Calculates the Fisher Information (expected information) of the given test, and returns a three dimensional array
of information matrices, where dimensions one and two run along the Q dimensions of the model, and three runs along items.

Fisher Information is given as;
\deqn{\mathcal{I}(\theta) = - \operatorname{E} \left[\left. \frac{\partial^2}{\partial\theta^2} \log f(X;\theta)\right|\theta \right]}{minus expectation of second derivative of the Log-Likelihood of f(theta)}

And is calculated as the weighted sum of second derivatives for all response categories. Information for multiple items is simply the sum of 
the individual information matrices.

Note: get_fisher_information always returns the 'raw' information, information given by prior distributions is added by the calling functions, 
and get_fisher_information(..) is normally called internally.
}

