% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_index_max_information_constrained.R
\name{get_item_index_max_information_constrained}
\alias{get_item_index_max_information_constrained}
\title{Shadow Test item selection}
\usage{
get_item_index_max_information_constrained(number_items, administered,
  available, answers, lp_constraints, lp_characters, item_information)
}
\arguments{
\item{number_items}{number of items in the test bank}

\item{administered}{vector with indices of administered items}

\item{available}{vector with indices of yet available items}

\item{answers}{vector with person answers}

\item{lp_constraints}{data frame with constraints in lp format: the lp_constraints from the list returned by constraints_lp_format()}

\item{lp_characters}{data frame with characteristics in lp format: the lp_chars from the list returned by constraints_lp_format()}

\item{item_information}{vector with summarized information of each yet available item, with zeros for administered items (returned by get_summarized_information() with pad = TRUE)}
}
\value{
integer item index of item with maximum information within constraints
}
\description{
Based on the Shadow Testing adaptive test assembly procedure by van der Linden (XXX)
}
\details{
At each step in a CAT, a Shadow Test is created consisting of an optimal complete test meeting all constraints.
The best item that was not already administrated is then selected from this test, and presented to the respondent.
}

