% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_information.R
\name{get_item_information}
\alias{get_item_information}
\title{Obtain a vector with information, summarized into one value, for each available item (values of the objective function).}
\usage{
get_item_information(information_summary, estimate, model, responses, prior,
  available, administered, number_items, number_dimensions, estimator, alpha,
  beta, guessing, number_itemsteps_per_item, lower_bound, upper_bound,
  pad = TRUE)
}
\arguments{
\item{information_summary}{called "objective" by Kroeze; how to summarize information; one of
"determinant": compute determinant(info_sofar_QxQ + info_QxQ_k) for each yet available item k
"posterior_determinant": compute determinant(info_sofar_QxQ_plus_prior + info_QxQ_k) for each yet available item k
"trace": compute trace((info_sofar_QxQ + info_QxQ_k) for each yet available item k
"posterior_trace": compute trace(info_sofar_QxQ_plus_prior + info_QxQ_k) for each yet available item k
"posterior_expected_kullback_leibler" = compute Posterior expected Kullback-Leibler Information}

\item{estimate}{vector with current theta estimate}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{responses}{vector with person responses}

\item{prior}{prior covariance matrix for theta}

\item{available}{vector with indeces of yet available items}

\item{administered}{vector with indeces of administered items}

\item{number_items}{number of items in test bank}

\item{number_dimensions}{number of dimensions}

\item{estimator}{type of estimator to be used, one of "maximum_aposteriori", "maximum_likelihood", or "expected_aposteriori"}

\item{alpha}{matrix containing the alpha parameters}

\item{beta}{matrix containing the beta parameters}

\item{guessing}{matrix containing the quessing parameters}

\item{number_itemsteps_per_item}{vector containing the number of non missing cells per row of the beta matrix}

\item{lower_bound}{vector with lower bounds for theta per dimension; estimated theta values smaller than the lowerbound values are truncated to the lowerbound values}

\item{upper_bound}{vector with upper bounds for theta per dimension; estimated theta values larger than the upperbound values are truncated to the upperbound values}

\item{pad}{Should the return vector be padded with zeros for items that have already been administered?}
}
\value{
vector with information for each available item
}
\description{
Information of the test so far (including all administered items) is added to the information of the available
item before the summary is computed
}

