% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_item.R
\name{get_next_item}
\alias{get_next_item}
\title{Next item}
\usage{
get_next_item(start_items, information_summary, lp_constraints, lp_characters,
  estimate, model, responses, prior, available, administered, number_items,
  number_dimensions, estimator, alpha, beta, guessing,
  number_itemsteps_per_item, lower_bound, upper_bound,
  eap_estimation_procedure = "riemannsum")
}
\arguments{
\item{start_items}{items that are shown to the patient before adaptive proces starts; one of
list(type = 'random', n)
list(type = 'fixed', indeces, n)
list(type = 'random_by_dimension', n_by_dimension, n)
where n = total number of initial items, indeces = vector of initial item indeces, 
n_by_dimension = scalar of number of initial items per dimension, or vector with number of initial items for each dimension}

\item{information_summary}{called "objective" by Kroeze; how to summarize information; one of
"determinant": compute determinant(info_sofar_QxQ + info_QxQ_k) for each yet available item k
"posterior_determinant": compute determinant(info_sofar_QxQ_plus_prior + info_QxQ_k) for each yet available item k
"trace": compute trace((info_sofar_QxQ + info_QxQ_k) for each yet available item k
"posterior_trace": compute trace(info_sofar_QxQ_plus_prior + info_QxQ_k) for each yet available item k
"posterior_expected_kullback_leibler" = compute Posterior expected Kullback-Leibler Information}

\item{lp_constraints}{data frame with constraints in lp format: the lp_constraints from the list returned by constraints_lp_format(); NULL means no constraints}

\item{lp_characters}{data frame with constraint characters in lp format: the lp_chars from the list returned by constraints_lp_format(); NULL means no constraints}

\item{estimate}{vector with current theta estimate}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{responses}{vector with person responses}

\item{prior}{covariance matrix of the (multi variate) normal prior for theta}

\item{available}{vector with indeces of yet available items}

\item{administered}{vector with indeces of administered items}

\item{number_items}{number of items in test bank}

\item{number_dimensions}{number of dimensions}

\item{estimator}{type of estimator to be used, one of "MAP" (Maximum a posteriori estimation), "EAP" (Expected A Posteriori Estimation), or "ML" (maximum likelihood)}

\item{alpha}{matrix containing the alpha parameters}

\item{beta}{matrix containing the beta parameters}

\item{guessing}{matrix containing the quessing parameters}

\item{number_itemsteps_per_item}{vector containing the number of non missing cells per row of the beta matrix}

\item{lower_bound}{vector with lower bounds for theta per dimension; estimated theta values smaller than the lowerbound values are truncated to the lowerbound values}

\item{upper_bound}{vector with upper bounds for theta per dimension; estimated theta values larger than the upperbound values are truncated to the upperbound values}

\item{eap_estimation_procedure}{String indicating the estimation procedure for the expected aposteriori estimate, which is computed
in get_posterior_expected_kl_information() if it is not the requested estimator in shadowcat(). One of "riemannsum" for integration via Riemannsum or
"gauss_hermite_quad" for integration via Gaussian Hermite Quadrature. Only important here if information_summary is posterior_expected_kl_information.}
}
\value{
integer item index next item
}
\description{
Control function to select the next item, takes account of starting conditions.
}

