% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_expected_kl_information.R
\name{get_posterior_expected_kl_information}
\alias{get_posterior_expected_kl_information}
\title{Posterior expected Kullback-Leibler Information}
\usage{
get_posterior_expected_kl_information(estimate, model, responses, administered,
  available, number_dimensions, estimator, alpha, beta, guessing, prior,
  number_itemsteps_per_item, lower_bound, upper_bound, theta_range = -3:3)
}
\arguments{
\item{estimate}{current theta estimate}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{responses}{vector with person responses}

\item{administered}{vector with indeces of administered items}

\item{available}{vector with indeces of available items}

\item{number_dimensions}{number of dimensions}

\item{estimator}{type of estimator to be used, one of "MAP" (Maximum a posteriori estimation), "EAP" (Expected A Posteriori Estimation), or "ML" (maximum likelihood)}

\item{alpha}{matrix containing the alpha parameters}

\item{beta}{matrix containing the beta parameters}

\item{guessing}{matrix containing the quessing}

\item{prior}{prior covariance matrix for theta; only required if estimator is "MAP" or "EAP" and output is "likelihoods" or "both"}

\item{number_itemsteps_per_item}{vector containing the number of non missing cells per row of the beta matrix}

\item{lower_bound}{vector with lower bounds for theta per dimension; estimated theta values smaller than the lowerbound values are truncated to the lowerbound values}

\item{upper_bound}{vector with upper bounds for theta per dimension; estimated theta values larger than the upperbound values are truncated to the upperbound values}

\item{theta_range}{Vector of theta values to be evaluated in the numerical integration. Using a sparser range may alleviate stress in higher dimensional tests.}
}
\value{
Vector with PEKL information for each yet available item.
}
\description{
Kullback-Leibler divergence based on the EAP and ML estimates of ability under the posterior distribution of theta.
Computes the numerical integral of the expectation of KL under the posterior distribution. The grid is simplified to a
range of -3, -2, ..., 3.
}
\details{
Note that even with a simplified grid, the number of quadrature points which have to be calculated for each available item, 
at each step in the CAT is taken to the power Q. Use of KL information is likely to be slow in 3+ dimensional tests. The theta range to be evaluated can be
specified to partially reduce the amount of computations required.
  
TODO: Add references.
}

