% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_expected_kl_information.R
\name{get_posterior_expected_kl_information}
\alias{get_posterior_expected_kl_information}
\title{Posterior expected Kullback-Leibler Information}
\usage{
get_posterior_expected_kl_information(estimate, model, answers, administered,
  available, number_dimensions, estimator, alpha, beta, guessing, prior_form,
  prior_parameters, number_itemsteps_per_item,
  eap_estimation_procedure = "riemannsum")
}
\arguments{
\item{estimate}{vector with current theta estimate}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{answers}{vector with person answers}

\item{administered}{vector with indices of administered items}

\item{available}{vector with indices of available items}

\item{number_dimensions}{number of dimensions}

\item{estimator}{type of estimator to be used, one of "maximum_aposteriori", "maximum_likelihood", or "expected_aposteriori"}

\item{alpha}{matrix containing the alpha parameters}

\item{beta}{matrix containing the beta parameters}

\item{guessing}{matrix containing the quessing parameters}

\item{prior_form}{String indicating the form of the prior; one of "normal" or "uniform"}

\item{prior_parameters}{List containing mu and Sigma of the normal prior: list(mu = ..., Sigma = ...), or 
the upper and lower bound of the uniform prior: list(lower_bound = ..., upper_bound = ...). Sigma should always
be in matrix form.}

\item{number_itemsteps_per_item}{vector containing the number of non missing cells per row of the beta matrix}

\item{eap_estimation_procedure}{String indicating the estimation procedure for the expected aposteriori estimate, which is computed
here if it is not the requested estimator in shadowcat(). One of "riemannsum" for integration via Riemannsum or
"gauss_hermite_quad" for integration via Gaussian Hermite Quadrature.}
}
\value{
Vector with PEKL information for each yet available item.
}
\description{
Kullback-Leibler divergence based on the EAP and ML estimates of ability under the posterior distribution of theta.
Computes the numerical integral of the expectation of KL under the posterior distribution. The grid is simplified to a
range of -3, -2, ..., 3.
}
\details{
Note that even with a simplified grid, the number of quadrature points which have to be calculated for each available item, 
at each step in the CAT is taken to the power Q. Use of KL information is likely to be slow in 3+ dimensional tests. The theta range to be evaluated can be
specified to partially reduce the amount of computations required.
  
TODO: Add references.
}

