% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probs_and_likelihoods_per_item.R
\name{get_probs_and_likelihoods_per_item}
\alias{get_probs_and_likelihoods_per_item}
\title{get list containing probabilities of scoring in each answer category for each item, and
likelihood with derivative per item}
\usage{
get_probs_and_likelihoods_per_item(theta, model, alpha, beta, guessing,
  answers = numeric(0), with_likelihoods)
}
\arguments{
\item{theta}{vector with true or estimated theta}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{alpha}{matrix containing the alpha parameters (if with_likelihoods, only of administered items)}

\item{beta}{matrix containing the beta parameters of administered items (if with_likelihoods, only of administered items)}

\item{guessing}{matrix containing the quessing parameters of administered items (if with_likelihoods, only of administered items)}

\item{answers}{if with_likelihoods = TRUE, vector with person answers to the administered items; else, numeric(0)}

\item{with_likelihoods}{if FALSE, only the probability matrix is returned, if TRUE, the likelihoods and derivatives are also returned}
}
\value{
list with P = matrix with for each included item (rows) the probability of scoring in each answer category (columns), given theta,
and if with_likelihoods, l = vector of likelihoods per item, d = vector of first derivatives, D = vector of second derivatives
}
\description{
get list containing probabilities of scoring in each answer category for each item, and
likelihood with derivative per item
}

