% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summarized_information.R
\name{get_summarized_information}
\alias{get_summarized_information}
\title{Obtain a vector with information, summarized into one value, for each available item (values of the objective function).}
\usage{
get_summarized_information(information_summary, estimate, model, answers,
  prior_form, prior_parameters, available, administered, number_items,
  number_dimensions, estimator, alpha, beta, guessing,
  number_itemsteps_per_item, pad = TRUE,
  eap_estimation_procedure = "riemannsum")
}
\arguments{
\item{information_summary}{called "objective" by Kroeze; how to summarize information; one of
"determinant": compute determinant(info_sofar_QxQ + info_QxQ_k) for each yet available item k
"posterior_determinant": compute determinant(info_sofar_QxQ_plus_prior_information + info_QxQ_k) for each yet available item k
"trace": compute trace((info_sofar_QxQ + info_QxQ_k) for each yet available item k
"posterior_trace": compute trace(info_sofar_QxQ_plus_prior_information + info_QxQ_k) for each yet available item k
"posterior_expected_kullback_leibler" = compute Posterior expected Kullback-Leibler Information}

\item{estimate}{vector with current theta estimate}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{answers}{vector with person answers}

\item{prior_form}{String indicating the form of the prior; one of "normal" or "uniform"}

\item{prior_parameters}{List containing mu and Sigma of the normal prior: list(mu = ..., Sigma = ...), or 
the upper and lower bound of the uniform prior: list(lower_bound = ..., upper_bound = ...). Sigma should always
be in matrix form.}

\item{available}{vector with indices of yet available items}

\item{administered}{vector with indices of administered items}

\item{number_items}{number of items in test bank}

\item{number_dimensions}{number of dimensions}

\item{estimator}{type of estimator to be used, one of "maximum_aposteriori", "maximum_likelihood", or "expected_aposteriori"}

\item{alpha}{matrix containing the alpha parameters}

\item{beta}{matrix containing the beta parameters}

\item{guessing}{matrix containing the quessing parameters}

\item{number_itemsteps_per_item}{vector containing the number of non missing cells per row of the beta matrix}

\item{pad}{Should the return vector be padded with zeros for items that have already been administered?}

\item{eap_estimation_procedure}{String indicating the estimation procedure for the expected aposteriori estimate, which is computed
in get_posterior_expected_kl_information() if it is not the requested estimator in shadowcat(). One of "riemannsum" for integration via Riemannsum or
"gauss_hermite_quad" for integration via Gaussian Hermite Quadrature. Only important here if information_summary is posterior_expected_kl_information.}
}
\value{
vector with information for each available item
}
\description{
Information of the test so far (including all administered items) is added to the information of the available
item before the summary is computed
}

