% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Not in}
\usage{
x \%not_in\% y
}
\arguments{
\item{x}{vector or value for which it is too be checked if its elements are in vector y}

\item{y}{vector for which it is too be checked if it contains the values in x}
}
\value{
Vector of same length as that of \code{x}, with elements equal to \code{TRUE} for each element of x that is not in y.
}
\description{
Get logical vector indicating if there no match for the left operant in the right operant.
}
\examples{
ShadowCAT:::`\%not_in\%`(c("a", "b", "g", "k", "b"), c("a", "b", "c"))
\dontshow{
all(ShadowCAT:::`\%not_in\%`(c("a", "b", "g", "k", "b"), c("a", "b", "c")) == c(FALSE, FALSE, TRUE, TRUE, FALSE)) || stop("wrong")
}
}

