% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itembankClass.R
\name{initItembank}
\alias{initItembank}
\title{Initialize an itembank object.}
\usage{
initItembank(model = "3PLM", alpha = NULL, beta = NULL, guessing = NULL,
  eta = NULL, silent = FALSE)
}
\arguments{
\item{model}{String, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{alpha}{Matrix of alpha paramteres, one column per dimension, one row per item. Note that so called within-dimensional models still use an alpha matrix, they simply
have only one non-zero loading per item.}

\item{beta}{Matrix of beta parameters, one column per item step, one row per item. Note that ShadowCAT expects response categories to be sequential,
and without gaps. That is, the weight parameter in the GPCM model is assumed to be sequential, and equal to the position of the 'location' of the beta parameter in the Beta matrix.
The matrix will have a number of columns equal to the largest number of response categories, items with fewer response categories should be
right-padded with \code{NA}. \code{NA} values between response categories are not allowed, and will lead to errors.
More flexibility in Beta parameters might be added in future versions.}

\item{guessing}{vector of guessing parameters per item. Optionally used in 3PLM model, ignored for all others.}

\item{eta}{Matrix of location parameters, optionally used in GPCM model, ignored for all others.}

\item{silent}{if TRUE, a summary of the item bank properties is printed}
}
\value{
ShadowCAT.items Itembank object, as used by \code{\link{initTest}} and \code{\link{initPerson}}.
}
\description{
Produce an itembank object from a list of parameters.
}
\details{
This is a convenience function to create an itembank object as expected by further ShadowCAT functions.
For all models, alpha is required (but may be set to a 1 vector/matrix for Rasch-type models).
Beta is required for all models, but in the case of GPCM, it will be computed from Eta if Eta is given and Beta omitted.
Guessing is only valid for the 3PLM model, and may safely be omitted (which implies a guessing parameter of 0 on all items).

Note that the GPCM expects Beta parameters to be category bounds rather than location parameters (Eta). That is, Beta_i = sum(Eta_1, ..., Eta_i).

TODO: model references.
}

