% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testClass.R
\name{initTest}
\alias{initTest}
\title{Initiate ShadowCAT test object}
\usage{
initTest(items, start = list(type = "random", n = 5), stop = list(type =
  "length", n = 30), max_n = 50, estimator = "MAP", objective = "PD",
  selection = "MI", constraints = NULL, lowerBound = rep(-3, items$Q),
  upperBound = rep(3, items$Q), ...)
}
\arguments{
\item{items}{}

\item{start}{}

\item{stop}{}

\item{max_n}{}

\item{estimator}{}

\item{objective}{}

\item{selection}{}

\item{constraints}{}

\item{lowerBound}{}

\item{upperBound}{}

\item{...}{}
}
\value{
ShadowCAT.test
}
\description{
Creates an object to be used with ShadowCAT
}

