% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testClass.R
\name{initTest}
\alias{initTest}
\title{Initiate ShadowCAT test object}
\usage{
initTest(items, start = NULL, stop = NULL, estimator = "MAP",
  objective = "PFI", selection = "MI", constraints = NULL, ...)
}
\arguments{
\item{items}{}

\item{start}{}

\item{stop}{}

\item{estimator}{}

\item{objective}{}

\item{selection}{}

\item{constraints}{}

\item{...}{}
}
\value{
ShadowCAT.test
}
\description{
Creates an object to be used with ShadowCAT
}

