% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{lapply_return_array}
\alias{lapply_return_array}
\title{lapply with array as output}
\usage{
lapply_return_array(x, dim, FUN, ...)
}
\arguments{
\item{x}{a vector or list; see ? lapply for details}

\item{dim}{vector with dimensions of the returned array}

\item{FUN}{function to be applied to each element of X; see ? lapply for details}

\item{...}{any additional arguments to FUN}
}
\value{
lapply output, with output converted to array
}
\description{
lapply with array as output
}
\examples{
matrix_example <- matrix(1:9, ncol = 3); 
lapply_return_array(1:3, 
                    c(3, 3, 3), 
                    FUN = function(i, matrix_example) { matrix_example[i,] \%*\% t(matrix_example[i,]) }, 
                    matrix_example = matrix_example)[1, 1, 1] == 1 || stop("wrong")
}

