% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_or_post_density.R
\name{likelihood_or_post_density}
\alias{likelihood_or_post_density}
\title{Get Likelihood or posterior density, and derivatives, based on a a given set of items}
\usage{
likelihood_or_post_density(theta, responses = NULL, model, items_to_include,
  number_dimensions, estimator, alpha, beta, guessing, prior = NULL,
  return_log_likelihood_or_post_density = TRUE,
  inverse_likelihood_or_post_density = FALSE)
}
\arguments{
\item{theta}{vector with true or estimated theta}

\item{responses}{vector with person responses to the administered items}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{items_to_include}{vector with indeces of items to which responses have been given}

\item{number_dimensions}{number of dimensions}

\item{estimator}{type of estimator to be used, one of "maximum_aposteriori", "maximum_likelihood", or "expected_aposteriori"}

\item{alpha}{matrix containing the alpha parameters (for complete test bank)}

\item{beta}{matrix containing the beta parameters (for complete test bank)}

\item{guessing}{matrix containing the quessing parameters (for complete test bank)}

\item{prior}{prior covariance matrix for theta; only required if estimator is "maximum_aposteriori" or "expected_aposteriori" and output is "likelihood" or "both"}

\item{return_log_likelihood_or_post_density}{if TRUE, log of likelihood or posterior density is returned, else likelihood or posterior density on original scale}

\item{inverse_likelihood_or_post_density}{should likelihood or posterior density values be reversed (useful for minimization, also reverses derivatives)}
}
\value{
the likelihood (estimator = maximum_likelihood) or posterior density (estimator = maximum_aposteriori or expected_aposteriori) of theta, with first and second derivatives as attributes
}
\description{
Get Likelihood or posterior density, and derivatives, based on a a given set of items
}

