% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_or_post_density.R
\name{likelihood_or_post_density}
\alias{likelihood_or_post_density}
\title{Get Likelihood or posterior density (with normal prior) over all included items, and derivatives, based on a a given set of items}
\usage{
likelihood_or_post_density(theta, answers = NULL, model, items_to_include,
  number_dimensions, estimator, alpha, beta, guessing,
  prior_parameters = NULL, return_log_likelihood_or_post_density = TRUE,
  inverse_likelihood_or_post_density = FALSE)
}
\arguments{
\item{theta}{vector with true or estimated theta}

\item{answers}{vector with person answers to the administered items}

\item{model}{string, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{items_to_include}{vector with indices of items to which answers have been given}

\item{number_dimensions}{number of dimensions}

\item{estimator}{type of estimator to be used, one of "maximum_aposteriori", "maximum_likelihood", or "expected_aposteriori"}

\item{alpha}{matrix containing the alpha parameters (for complete test bank)}

\item{beta}{matrix containing the beta parameters (for complete test bank)}

\item{guessing}{matrix containing the quessing parameters (for complete test bank)}

\item{prior_parameters}{List containing mu and Sigma of the normal prior: list(mu = ..., Sigma = ...).
Sigma should always be in matrix form.}

\item{return_log_likelihood_or_post_density}{if TRUE, log of likelihood or posterior density is returned, else likelihood or posterior density on original scale}

\item{inverse_likelihood_or_post_density}{should likelihood or posterior density values be reversed (useful for minimization, also reverses derivatives)}
}
\value{
the likelihood (estimator is maximum_likelihood) or posterior density with normal prior (estimator is not maximum_likelihood) of theta, with first and second derivatives as attributes
}
\description{
Get Likelihood or posterior density (with normal prior) over all included items, and derivatives, based on a a given set of items
}

