% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prob.R
\name{LL}
\alias{LL}
\title{Log Likelihood}
\usage{
LL(theta, test, person, minimize = FALSE)
}
\arguments{
\item{theta}{}

\item{test}{}

\item{person}{}

\item{should}{values be reversed (useful for minimization, reverses LL as well as derivatives)}
}
\value{
Log-Likelihood, as well as gradient and hessian attributes.
}
\description{
Internal - fetch appropriate elements from prob for nlm optimizer
}

