% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{matrix_apply}
\alias{matrix_apply}
\title{apply with input and output converted to matrix}
\usage{
matrix_apply(x, margin, FUN, ...)
}
\arguments{
\item{x}{a vector or matrix; if vector, the vector will be converted to a matrix}

\item{margin}{a vector giving the subscripts which the function will be applied over, see ?apply for details}

\item{FUN}{the funcion to be applied}

\item{...}{any additional arguments to FUN}
}
\value{
same a apply, but always in matrix format
}
\description{
apply with input and output converted to matrix
}
\examples{
( matrix_apply(c(1:5, NA), 2, sum, na.rm = TRUE) == 15 && is.matrix(matrix_apply(c(1:5, NA), 2, sum, na.rm = TRUE)) ) || stop("wrong")
}

