% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/selection_MI.R
\name{MI}
\alias{MI}
\title{Maximum Information item selection}
\usage{
MI(test, person, objective)
}
\arguments{
\item{test}{}

\item{person}{}
}
\value{
item index of item with the highest value of the objective function.
}
\description{
Naive item selection based on maximum information only.
}
\details{
Selects the item with the highest value of a specified objective function.
}

