% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{na_only_end_rows}
\alias{na_only_end_rows}
\title{Check whether NA's exist only at the end of each row, that is, whether there are no values after the first NA in the row}
\usage{
na_only_end_rows(x)
}
\arguments{
\item{x}{Matrix}
}
\value{
TRUE if there are only NA's at the end of each row or no NA at all; FALSE otherwise
}
\description{
Check whether NA's exist only at the end of each row, that is, whether there are no values after the first NA in the row
}
\examples{
na_only_end_rows(matrix(c(1:3, NA, NA, 1:5, 1:4, NA), ncol = 5, byrow = TRUE)) || stop("wrong");
na_only_end_rows(matrix(c(1:3, NA, NA, 1:2, NA, 4:5, 1:4, NA), ncol = 5, byrow = TRUE))  == FALSE || stop("wrong")
}

