% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{na_only_end_vector}
\alias{na_only_end_vector}
\title{Check whether NA's exist only at the end of a vector, that is, whether there are no values after the first NA in the vector}
\usage{
na_only_end_vector(x)
}
\arguments{
\item{x}{Vector}
}
\value{
TRUE if there are only NA's at the end of the vector or no NA at all; FALSE otherwise
}
\description{
Check whether NA's exist only at the end of a vector, that is, whether there are no values after the first NA in the vector
}
\examples{
na_only_end_vector(c(1:3, NA, NA)) || stop("wrong");
na_only_end_vector(c(1:3, NA, NA, 6)) == FALSE || stop("wrong")
}

