% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/selection_MI.R
\name{next_item}
\alias{next_item}
\title{Next item in an adaptive test. Takes a person and test object, and returns the index of the next item based}
\usage{
next_item(test, person)
}
\arguments{
\item{test}{}

\item{person}{}
}
\value{
integer item index
}
\description{
This function is a wrapper that sends the actual work to the (hopefully) correct subroutines.
}

