% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{no_missing_information}
\alias{no_missing_information}
\title{Check whether characteristics and constraints are either both NULL or both not NULL}
\usage{
no_missing_information(characteristics, constraints)
}
\arguments{
\item{characteristics}{Data frame with characteristics.}

\item{constraints}{List of constraints.}
}
\value{
TRUE if characteristics and constraints are either both NULL or both not NULL, FALSE otherwise
}
\description{
Check whether characteristics and constraints are either both NULL or both not NULL
}

