% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{number_non_missing_cells_per_row}
\alias{number_non_missing_cells_per_row}
\title{Count non-\code{NA} per row}
\usage{
number_non_missing_cells_per_row(x)
}
\arguments{
\item{x}{Matrix.}
}
\value{
Vector with the number of non-\code{NA} cells per row.
}
\description{
Get the number of non-\code{NA} cells per row of a matrix.
}
\examples{
ShadowCAT:::number_non_missing_cells_per_row(matrix(c(1,2,NA,NA,2,3,NA,2,3), ncol = 3))

\dontshow{
all(ShadowCAT:::number_non_missing_cells_per_row(matrix(c(1,2,NA,NA,2,3,NA,2,3), ncol = 3)) == c(1, 3, 2)) || stop("wrong")
}
}

