% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_expected_kl_information.R
\name{PEKL}
\alias{PEKL}
\title{Posterior expected Kullback-Leibler Information}
\usage{
PEKL(test, person, theta_range = -3:3)
}
\arguments{
\item{test}{Test object, see \code{\link{initTest}}.}

\item{person}{Person object, see \code{\link{initPerson}}.}

\item{theta_range}{Vector of theta values to be evaluated in the numerical integration. Using a sparser range may alleviate stress in higher dimensional tests.}
}
\value{
Vector with PEKL information for each item.
}
\description{
Kullback-Leibler divergence based on the EAP and ML estimates of ability under the posterior distribution of theta.
Computes the numerical integral of the expectation of KL under the posterior distribution. The grid is simplified to a
range of -3, -2, ..., 3.
}
\details{
Note that even with a simplified grid, the number of quadrature points which have to be calculated for each available item, 
at each step in the CAT is taken to the power Q. Use of KL information is likely to be slow in 3+ dimensional tests. The theta range to be evaluated can be
specified to partially reduce the amount of computations required.
  
TODO: Add references.
}

