% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob}
\alias{prob}
\title{Probability and Derivatives for a given (subset of) ShadowCAT.items itembank. It gives the probability of 
responses given theta (only useful for simulation I think), and the likelihood of given responses}
\usage{
prob(test, person = NULL, theta = NULL, deriv = FALSE, prior = NULL,
  items = NULL)
}
\arguments{
\item{test}{test object}

\item{person}{Person object, will use current theta estimate of the person to obtain P values.}

\item{theta}{Give a specific theta vector to use. Overrides person$estimate if set.}

\item{deriv}{Should we fetch derivatives and LL?}

\item{prior}{prior to be applied to derivatives. Overrides person$prior if set}

\item{items}{overrides test$items if set}
}
\description{
Generic
}
\details{
Details
}

