% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prob.R
\name{prob}
\alias{prob}
\title{Probability and Derivatives for a given (subset of) ShadowCAT.items itembank.}
\usage{
prob(test, person = NULL, theta = NULL, deriv = FALSE, prior = NULL)
}
\arguments{
\item{person}{Person object, will use current theta estimate of the person to obtain P values.}

\item{theta}{Give a specific theta vector to use, overrides person if set.}

\item{deriv}{Should we fetch derivatives and LL?}

\item{prior}{If not NULL, prior to be applied to derivatives.}

\item{items}{}
}
\description{
Generic
}
\details{
Details
}

