% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{row_cumsum}
\alias{row_cumsum}
\title{compute cumulative sums for each row of a matrix}
\usage{
row_cumsum(X)
}
\arguments{
\item{X}{a vector or matrix; if vector, the vector will be converted to a matrix with one column}
}
\value{
matrix containing cumulative sums for each row
}
\description{
compute cumulative sums for each row of a matrix
}
\examples{
row_cumsum(matrix(1:12, ncol = 3))[2,] == c(2, 6, 10) || stop("wrong");
row_cumsum(1:12) == matrix(1:12, ncol = 1) || stop("wrong")
}

