% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{row_cumsum}
\alias{row_cumsum}
\title{Cumulative row sums}
\usage{
row_cumsum(x)
}
\arguments{
\item{x}{Vector or matrix. If vector, the vector will be converted to a matrix with one column.}
}
\value{
Matrix containing cumulative sums for each row.
}
\description{
Compute cumulative sums for each row of a matrix.
}
\examples{
ShadowCAT:::row_cumsum(matrix(1:12, ncol = 3))
ShadowCAT:::row_cumsum(1:12)
ShadowCAT:::row_cumsum(matrix(1:4, nrow = 1))

\dontshow{
all(ShadowCAT:::row_cumsum(matrix(1:12, ncol = 3))[2,] == c(2, 8, 18)) || stop("wrong")
all(should_be_equal2 <- ShadowCAT:::row_cumsum(1:12) == matrix(1:12, ncol = 1)) || stop("wrong")
all(should_be_equal3 <- ShadowCAT:::row_cumsum(matrix(1:4, nrow = 1)) == matrix(c(1, 3, 6, 10), nrow = 1)) || stop("wrong")
}
}

