% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{rsapply}
\alias{rsapply}
\title{Recursive list apply, given a bunch of vectors, creates lists of lists with the elements as keys
 and values determined by fn.}
\usage{
rsapply(..., fn)
}
\arguments{
\item{...}{vectors to loop over}

\item{fn}{function to execute for each leaf}
}
\description{
Recursive list apply, given a bunch of vectors, creates lists of lists with the elements as keys
 and values determined by fn.
}
\examples{
rsapply(c('a', 'b'), c('f', 'g'), fn=function(x,y) c(x, y))
#> list(a=list(f=c('a', 'f'), g=c('a', 'g')), b=list(f=c('b', 'f'), g=c('b', 'g')))
}

