% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ShadowCAT.R
\name{ShadowCAT}
\alias{ShadowCAT}
\title{ShadowCAT}
\usage{
ShadowCAT(person, test, verbose = "argh!")
}
\arguments{
\item{person}{}

\item{test}{}
}
\value{
person
}
\description{
Run test with a specified person, and a specified test. See initPerson and initTest.
 This is a simple wrapper to call the right methods, options should be defined in the test object.
}
\details{
Details
}

