% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functions.R
\name{simulate_answer}
\alias{simulate_answer}
\title{Simulates answer}
\usage{
simulate_answer(theta, model, alpha, beta, guessing, item_keys)
}
\arguments{
\item{theta}{Vector with true theta.}

\item{model}{String, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{alpha}{Matrix of alpha parameters. See \code{shadowcat} for details.}

\item{beta}{Matrix of beta parameters, one column per item step, one row per item. See \code{shadowcat} for details.}

\item{guessing}{One column matrix of guessing parameters per item. Row names should contain the item keys. Optionally used in 3PLM model, ignored for all others.}

\item{item_keys}{Character vector of item keys for which answers should be simulated.}
}
\value{
Vector with responses
}
\description{
Simulates answer on specified items, given true theta. Only for testing purposes
}
\examples{
alpha_beta <- simulate_testbank(model = "3PLM", number_items = 50, number_dimensions = 1, number_itemsteps = 1)
guessing <- matrix(rep(.5, 50), dimnames = list(rownames(alpha_beta$alpha), NULL))

# Without guessing parameter
simulate_answer(theta = .3, model = "3PLM", alpha = alpha_beta$alpha, beta = alpha_beta$beta, guessing = NULL, item_keys = "item3")

# With guessing parameter
simulate_answer(theta = .3, model = "3PLM", alpha = alpha_beta$alpha, beta = alpha_beta$beta, guessing = guessing, item_keys = "item3")

# Simulate answers for more than one item
simulate_answer(theta = .3, model = "3PLM", alpha = alpha_beta$alpha, beta = alpha_beta$beta, guessing = NULL, item_keys = c("item5", "item2", "item8", "item1", "item18"))
}

