% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_testbank.R
\name{simulate_testbank}
\alias{simulate_testbank}
\title{Create a test Itembank}
\usage{
simulate_testbank(model, K = 50, Q = 1, M = 4, between = FALSE,
  run_initItembank = TRUE)
}
\arguments{
\item{model}{String, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{K}{number of items}

\item{Q}{number of dimensions}

\item{M}{number of item steps (number of categories minus 1)}

\item{between}{is TRUE, force items to load on one dimension each}

\item{run_initItembank}{if FALSE, the simulated testbank is returned; if TRUE, initItembank() is applied on
the simulated testbank and the result of this is returned}
}
\value{
ShadowCAT.itembank
}
\description{
Quick and simple itembanks for testing purposes.
}

