% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset_shadowcat_items.R
\name{subset.ShadowCAT.items}
\alias{subset.ShadowCAT.items}
\title{Subset method for ShadowCAT.items.}
\usage{
\method{subset}{ShadowCAT.items}(items, select)
}
\arguments{
\item{items}{item characteristics class ShadowCAT.items.}

\item{select}{indeces of items to be selected}
}
\value{
Item subset of provided itembank, with class ShadowCAT.items, as well as any original classes.
}
\description{
Creates a valid subset of items to work with in ShadowCAT functions. Primarily used internally.
}

