% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itembankClass.R
\name{subset.ShadowCAT.items}
\alias{subset.ShadowCAT.items}
\title{Subset method for ShadowCAT.items.}
\usage{
\method{subset}{ShadowCAT.items}(items, subset)
}
\arguments{
\item{items}{class ShadowCAT.items.}

\item{subset}{indeces.}
}
\value{
items Subset of provided itembank, with class ShadowCAT.items, as well as any original classes.
}
\description{
Creates a valid subset of items to work with in ShadowCAT functions. Primarily used internally.
}

