% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_shadowcat_functions.R
\name{sum_loop_outputs}
\alias{sum_loop_outputs}
\title{Sum loop outputs}
\usage{
sum_loop_outputs(start_object, loop_vector, FUN, ...)
}
\arguments{
\item{start_object}{The object to which the sum is to be added.}

\item{loop_vector}{The vector to be looped over.}

\item{FUN}{Function of the loop_vector values, returning the object to be added to \code{start_object} at each iteration. First argument should be loop values argument.}

\item{...}{Any additional arguments to FUN.}
}
\value{
Sum of objects returned by a loop, added to starting object.
}
\description{
Get sum of objects returned by a loop, added to a starting object.
}
\examples{
matrix_example <- matrix(1:6, ncol = 2)
ShadowCAT:::sum_loop_outputs(start_object = matrix(0, 2, 2), loop_vector = 1:3, 
                             FUN = function(item, matrix_example) { 
                                     matrix_example[item,] \%*\% t(matrix_example[item,]) }, 
                             matrix_example = matrix_example)

 \dontshow{
all(ShadowCAT:::sum_loop_outputs(start_object = matrix(0, 2, 2), loop_vector = 1:3, FUN = function(item, matrix_example) { matrix_example[item,] \%*\% t(matrix_example[item,]) }, matrix_example = matrix_example) == 
                   (matrix_example[1,] \%*\% t(matrix_example[1,]) + matrix_example[2,] \%*\% t(matrix_example[2,]) + matrix_example[3,] \%*\% t(matrix_example[3,]))) || stop("wrong")
}
}

