% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functions.R
\name{test_shadowcat}
\alias{test_shadowcat}
\title{Simulate a testing routine with shadowcat}
\usage{
test_shadowcat(true_theta, prior_form, prior_parameters, model, alpha, beta,
  guessing, eta = NULL, start_items, stop_test, estimator,
  information_summary, constraints_and_characts = NULL, lower_bound = NULL,
  upper_bound = NULL, safe_eap = FALSE, initital_estimate = rep(0,
  ncol(alpha)), initial_variance = diag(ncol(alpha)) * 25,
  eap_estimation_procedure = "riemannsum")
}
\arguments{
\item{true_theta}{True theta value or vector}

\item{prior_form}{String indicating the form of the prior; one of "normal" or "uniform"}

\item{prior_parameters}{List containing mu and Sigma of the normal prior: list(mu = ..., Sigma = ...), or 
the upper and lower bound of the uniform prior: list(lower_bound = ..., upper_bound = ...). Sigma should always
be in matrix form. The length of lower_bound and upper_bound should be equal to the number of dimensions}

\item{model}{String, one of '3PLM', 'GPCM', 'SM' or 'GRM', for the three-parameter logistic, generalized partial credit, sequential or graded response model respectively.}

\item{alpha}{Matrix of alpha parameters. See \code{shadowcat} for details.}

\item{beta}{Matrix of beta parameters. See \code{shadowcat} for details.}

\item{guessing}{Matrix with one column of guessing parameters per item. See \code{shadowcat} for details.}

\item{eta}{Matrix of location parameters, optionally used in GPCM model, ignored for all others. See \code{shadowcat} for details.}

\item{start_items}{List indicating the items that are shown to the respondent before adaptive proces starts. One of
list(type = 'random', n)
list(type = 'fixed', indices, n)
list(type = 'random_by_dimension', n_by_dimension, n)
See \code{shadowcat} for details.}

\item{stop_test}{List indicating rules for when to stop providing new items to respondent. Should be of the form
list(target = ..., max_n = ..., min_n = ..., cutoffs = ...). See \code{shadowcat} for details.}

\item{estimator}{Type of estimator to be used, one of "maximum_aposteriori", "maximum_likelihood", or "expected_aposteriori".}

\item{information_summary}{How to summarize Fisher information, used for selection of item with maximum information. 
One of "determinant", "posterior_determinant", "trace", "posterior_trace", or "posterior_expected_kullback_leibler". See \code{shadowcat} for details.}

\item{constraints_and_characts}{list with constraints and characteristics. See \code{shadowcat} for details.}

\item{lower_bound}{Vector with lower bounds for theta per dimension; estimated theta values smaller than the lowerbound values are truncated to the lowerbound values.
Can only be defined when estimator is maximum_likelihood. Setting bounds with maximum likelihood estimation is equivalent to
using maximum aposteriori estimation with a uniform prior.}

\item{upper_bound}{Vector with upper bounds for theta per dimension; estimated theta values larger than the upperbound values are truncated to the upperbound values
Can only be defined when estimator is maximum_likelihood. Setting bounds with maximum likelihood estimation is equivalent to
using maximum aposteriori estimation with a uniform prior.}

\item{safe_eap}{Only relevant if estimator is expected_aposteriori. 
TRUE if estimator should switch to maximum aposteriori if the integration algorithm results in an error.
An error may occur if the prior is uniform, estimator is expected aposteriori, and the bounds do not exceed the true theta value, or are too close to it.}

\item{initital_estimate}{Vector containing the initial theta estimates (starting values)}

\item{initial_variance}{Matrix containing the initial covariance matrix (staring values)}

\item{eap_estimation_procedure}{String indicating the estimation procedure if estimator is expected aposteriori. One of "riemannsum" for integration via Riemannsum or
"gauss_hermite_quad" for integration via Gaussian Hermite Quadrature.}
}
\value{
List as returned by shadowcat after the last answer, with variance element turned into matrix
}
\description{
Simulate a testing routine with shadowcat
}
\examples{
# One dimension
alpha_beta_one_dim <- simulate_testbank(model = "GPCM", number_items = 50, number_dimensions = 1, number_itemsteps = 3)
test_shadowcat(true_theta = 2, prior_form = "normal", prior_parameters = list(mu = 0, Sigma = diag(1)), model = "SM", alpha = alpha_beta_one_dim$alpha, beta = alpha_beta_one_dim$beta, guessing = NULL, start_items = list(type = 'random', n = 3), stop_test = list(max_n = 20, target = 0.1), estimator = "maximum_aposteriori", information_summary = "posterior_determinant")

# Three dimensions
alpha_beta_three_dim <- simulate_testbank(model = "GPCM", number_items = 100, number_dimensions = 3, number_itemsteps = 3)
test_shadowcat(true_theta = c(0, 1, -.5), prior_form = "normal", prior_parameters = list(mu = c(0, 0, 0), Sigma = diag(3)), model = "SM", alpha = alpha_beta_three_dim$alpha, beta = alpha_beta_three_dim$beta, guessing = NULL, start_items = list(type = 'random_by_dimension', n_by_dimension = 3, n = 9), stop_test = list(max_n = 60, target = c(.1, .1, .1)), estimator = "maximum_aposteriori", information_summary = "posterior_determinant")
}

