% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_and_run.R
\name{validate_and_run}
\alias{validate_and_run}
\title{Validate and run}
\usage{
validate_and_run()
}
\description{
Validate input and run calling method. Template to structure and test larger methods.
}
\details{
First \code{validate} is called. If any errors are added, the value of \code{invalid_result} is returned, 
else the value of \code{result}.
Loads the following methods into the environment of the calling method:
\itemize{
\item \code{add_error(key, value)}
\item \code{errors()} #returns \code{list(key=value, key2=value2)}
\item \code{validate_and_runner()} # Used internally, contains the logic of functions to call.}
}
\examples{
two_times_two <- function(two, also_two) {
  result <- function() helper()

  helper <- function() two * also_two

  validate <- function() {
    if(two != 2) add_error('two', 'should be 2')
    if(also_two != 2) add_error('also_two', 'should also be two')
  }

  invalid_result <- function() {
    list(result=NA, errors=errors())
  }

  ShadowCAT:::validate_and_run()
}

# No errors
two_times_two(2,2)

# Error
two_times_two(3,2)

\dontshow{
two_times_two(2,2) == 4 || stop('not 4')
two_times_two(3,2)$errors$two == 'should be 2' || stop('wrong result')
}
}

